fix(guard): review uninspectable interpreter stdin
