--- .github/pr2870-finalize/000.b64
@@ -1,1 +1,0 @@
-ZGlmZiAtLWdpdCBhL3NyYy9jb2RleF9wbHVnaW5fc2Nhbm5lci9ndWFyZC9ydW50aW1lL19zaGVsbF9zZWNyZXRfcmVhZF9zdXBwb3J0LnB5IGIvc3JjL2NvZGV4X3BsdWdpbl9zY2FubmVyL2d1YXJkL3J1bnRpbWUvX3NoZWxsX3NlY3JldF9yZWFkX3N1cHBvcnQucHkKbmV3IGZpbGUgbW9kZSAxMDA2NDQKaW5kZXggMDAwMDAwMC4uYzkyMmM4OAotLS0gL2Rldi9udWxsCisrKyBiL3NyYy9jb2RleF9wbHVnaW5fc2Nhbm5lci9ndWFyZC9ydW50aW1lL19zaGVsbF9zZWNyZXRfcmVhZF9zdXBwb3J0LnB5CkBAIC0wLDAgKzEsNDcwIEBACisiIiJQcml2YXRlIHBhcnNpbmcgYW5kIHBhdGggaGVscGVycyBmb3Igc2hlbGwgc2VjcmV0LXJlYWQgYXNzZXNzbWVudC4iIiIKKworZnJvbSBfX2Z1dHVyZV9fIGltcG9ydCBhbm5vdGF0aW9ucworCitpbXBvcnQgcmUKK2Zyb20gZGF0YWNsYXNzZXMgaW1wb3J0IHJlcGxhY2UKK2Zyb20gcGF0aGxpYiBpbXBvcnQgUGF0aAorCitmcm9tIC5fc2hlbGxfZXhlY3V0aW9uX2NvbnRleHRfc3VwcG9ydCBpbXBvcnQgKAorICAgIFNIRUxMX0NXRF9NSVNTSU5HX0RJUkVDVE9SWSwKKyAgICBTSEVMTF9DV0RfTk9UX0RJUkVDVE9SWSwKKyAgICBTSEVMTF9DV0RfVU5SRUFEQUJMRV9ESVJFQ1RPUlksCisgICAgc3BsaXRfc2hlbGxfdG9rZW5zLAorKQorZnJvbSAuY29tbWFuZF9tb2RlbCBpbXBvcnQgQ2Fub25pY2FsQ29tbWFuZCwgQ29tbWFuZFNlZ21lbnQKK2Zyb20gLmhvbWVfcGF0aF90ZXh0IGltcG9ydCBleHBhbmRfaG9tZSwgbm9ybWFsaXplX3BhdGgKK2Zyb20gLnNlY3JldF9zZW5zaXRpdml0eSBpbXBvcnQgY2xhc3NpZnlfc2VjcmV0X3BhdGgKK2Zyb20gLnNoZWxsX2V4ZWN1dGlvbl9jb250ZXh0IGltcG9ydCBTaGVsbEV4ZWN1dGlvblNlZ21lbnQKKworX1NIRUxMUyA9IGZyb3plbnNldCh7InNoIiwgImJhc2giLCAiZGFzaCIsICJhc2giLCAienNoIiwgImtzaCIsICJmaXNoIiwgInNvdXJjZSIsICIuIn0pCitfU0NSSVBUX1NVRkZJWEVTID0gKAorICAgICIuc2giLAorICAgICIuYmFzaCIsCisgICAgIi56c2giLAorICAgICIua3NoIiwKKyAgICAiLmZpc2giLAorICAgICIucHkiLAorICAgICIuanMiLAorICAgICIubWpzIiwKKyAgICAiLmNqcyIsCisgICAgIi50cyIsCisgICAgIi5yYiIsCisgICAgIi5wbCIsCispCitfT1RIRVJfUkVBREVSUyA9IGZyb3plbnNldCh7ImJhc2U2NCIsICJ4eGQiLCAib2QiLCAiaGV4ZHVtcCIsICJzdHJpbmdzIiwgInRhYyIsICJsZXNzIiwgIm1vcmUiLCAic29ydCIsICJ1bmlxIiwgIndjIn0pCitfTUFYX1NDUklQVFMgPSAxNgorX01BWF9ERVBUSCA9IDQKK19NQVhfVE9UQUxfQllURVMgPSAxMjggKiAxMDI0CitfTUFYX0lOTElORV9TQ1JJUFRfQllURVMgPSA2NCAqIDEwMjQKK19QWVRIT05fRVhFQ1VUQUJMRSA9IHJlLmNvbXBpbGUociJweXRob253Pyg/OlxkKyg/OlwuXGQrKSopPyg/OlwuZXhlKT8kIiwgcmUuSUdOT1JFQ0FTRSkKK19MSVRFUkFMX1JFQUQgPSByZS5jb21waWxlKAorICAgIHIiKD86XGJvcGVufFxicmVhZEZpbGUoPzpTeW5jKT98XGJjcmVhdGVSZWFkU3RyZWFtfFxiQnVuXC5maWxlfFxibG9hZF9kb3RlbnYpIgorICAgIHIiXHMqXChccyooWydcIl0pKFteJ1wiXG5ceDAwXXsxLDQwOTZ9KVwxIgorKQorX1BBVEhfUkVBRCA9IHJlLmNvbXBpbGUoCisgICAgciJcYlBhdGhccypcKFxzKihbJ1wiXSkoW14nXCJcblx4MDBdezEsNDA5Nn0pXDFccypcKSIKKyAgICByIlxzKlwuXHMqKD86cmVhZF90ZXh0fHJlYWRfYnl0ZXN8b3BlbilccypcKCIKKykKK19TSE9SVF9DSVJDVUlUSU5HX0NEX0ZBSUxVUkVTID0gZnJvemVuc2V0KAorICAgIHtTSEVMTF9DV0RfTUlTU0lOR19ESVJFQ1RPUlksIFNIRUxMX0NXRF9OT1RfRElSRUNUT1JZLCBTSEVMTF9DV0RfVU5SRUFEQUJMRV9ESVJFQ1RPUll9CispCitfRkxPV19PUEVSQVRPUlMgPSBmcm96ZW5zZXQoeyImJiIsICJ8fCIsICJ8IiwgIjsiLCAiJiJ9KQorCisKK2RlZiBfbGl0ZXJhbF9yZWFkX3BhdGhzKHNvdXJjZTogc3RyKSAtPiB0dXBsZVtzdHIsIC4uLl06CisgICAgIiIiUmV0dXJuIGJvdW5kZWQgbGl0ZXJhbCBmaWxlIG9wZXJhbmRzIGZyb20gc3VwcG9ydGVkIGlubGluZSBydW50aW1lcy4iIiIKKworICAgIHJldHVybiB0dXBsZShtYXRjaC5ncm91cCgyKSBmb3IgcGF0dGVybiBpbiAoX0xJVEVSQUxfUkVBRCwgX1BBVEhfUkVBRCkgZm9yIG1hdGNoIGluIHBhdHRlcm4uZmluZGl0ZXIoc291cmNlKSkKKworCitkZWYgX3B5dGhvbl9leGVjdXRhYmxlKG5hbWU6IHN0cikgLT4gYm9vbDoKKyAgICByZXR1cm4gX1BZVEhPTl9FWEVDVVRBQkxFLmZ1bGxtYXRjaChuYW1lKSBpcyBub3QgTm9uZQorCisKK2RlZiBfY29tbWFuZF9tYXlfbmVlZF9yZWFkX2Fzc2Vzc21lbnQoY29tbWFuZF90ZXh0OiBzdHIpIC0+IGJvb2w6CisgICAgIiIiU2tpcCBmaWxlc3lzdGVtIG1vZGVsaW5nIHdoZW4gc3ludGF4IGNhbm5vdCByZWFkIGZpbGVzIG9yIGxhdW5jaCBsb2NhbCBjb2RlLiIiIgorCisgICAgdHJ5OgorICAgICAgICB0b2tlbnMgPSBzcGxpdF9zaGVsbF90b2tlbnMoY29tbWFuZF90ZXh0KQorICAgIGV4Y2VwdCBWYWx1ZUVycm9yOgorICAgICAgICByZXR1cm4gVHJ1ZQorICAgIGludGVyZXN0aW5nID0geworICAgICAgICAqX1NIRUxMUywKKyAgICAgICAgKl9PVEhFUl9SRUFERVJTLAorICAgICAgICAiY2F0IiwKKyAgICAgICAgImhlYWQiLAorICAgICAgICAidGFpbCIsCisgICAgICAgICJzZWQiLAorICAgICAgICAiZ3JlcCIsCisgICAgICAgICJlZ3JlcCIsCisgICAgICAgICJmZ3JlcCIsCisgICAgICAgICJyZyIsCisgICAgICAgICJyZWFkIiwKKyAgICAgICAgImNvbW1hbmQiLAorICAgICAgICAiZXhlYyIsCisgICAgICAgICJub2RlIiwKKyAgICAgICAgImJ1biIsCisgICAgICAgICJydWJ5IiwKKyAgICAgICAgInBlcmwiLAorICAgIH0KKyAgICBmb3IgdG9rZW4gaW4gdG9rZW5zOgorICAgICAgICBpZiAiPCIgaW4gdG9rZW46CisgICAg
