--- .github/pr2870-finalize/001.b64
@@ -1,1 +1,0 @@
-ICAgICAgICByZXR1cm4gVHJ1ZQorICAgICAgICBub3JtYWxpemVkID0gdG9rZW4uc3RyaXAoKQorICAgICAgICBpZiBub3Qgbm9ybWFsaXplZDoKKyAgICAgICAgICAgIGNvbnRpbnVlCisgICAgICAgIG5hbWUgPSAiLiIgaWYgbm9ybWFsaXplZCA9PSAiLiIgZWxzZSBQYXRoKG5vcm1hbGl6ZWQpLm5hbWUubG93ZXIoKQorICAgICAgICBpZiBuYW1lIGluIGludGVyZXN0aW5nIG9yIF9weXRob25fZXhlY3V0YWJsZShuYW1lKToKKyAgICAgICAgICAgIHJldHVybiBUcnVlCisgICAgICAgIGlmIF9wYXRoX3F1YWxpZmllZChub3JtYWxpemVkKToKKyAgICAgICAgICAgIHJldHVybiBUcnVlCisgICAgcmV0dXJuIEZhbHNlCisKKworZGVmIF9zZW5zaXRpdmVfcGF0aCh2YWx1ZTogc3RyLCAqLCBjd2Q6IFBhdGggfCBOb25lLCBob21lX2RpcjogUGF0aCB8IE5vbmUpIC0+IHN0ciB8IE5vbmU6CisgICAgIiIiQ2xhc3NpZnkgYSBsaXRlcmFsIG9yIG1ldGFkYXRhLXJlc29sdmVkIGFsaWFzIHdpdGhvdXQgb3BlbmluZyBzZWNyZXRzLiIiIgorCisgICAgbWF0Y2ggPSBjbGFzc2lmeV9zZWNyZXRfcGF0aCh2YWx1ZSwgY3dkPWN3ZCwgaG9tZV9kaXI9aG9tZV9kaXIpCisgICAgaWYgbWF0Y2ggaXMgbm90IE5vbmU6CisgICAgICAgIHJldHVybiBtYXRjaC5wYXRoCisgICAgbGV4aWNhbCA9IFBhdGgobm9ybWFsaXplX3BhdGgoZXhwYW5kX2hvbWUodmFsdWUsIGhvbWVfZGlyKSwgY3dkKSkKKyAgICByb290cyA9IHR1cGxlKHJvb3QgZm9yIHJvb3QgaW4gKGN3ZCwgaG9tZV9kaXIpIGlmIHJvb3QgaXMgbm90IE5vbmUpCisgICAgaWYgbm90IGxleGljYWwuaXNfYWJzb2x1dGUoKSBvciBub3QgYW55KGxleGljYWwuaXNfcmVsYXRpdmVfdG8ocm9vdCkgZm9yIHJvb3QgaW4gcm9vdHMpOgorICAgICAgICByZXR1cm4gTm9uZQorICAgIHRyeToKKyAgICAgICAgcmVzb2x2ZWQgPSBsZXhpY2FsLnJlc29sdmUoc3RyaWN0PVRydWUpCisgICAgZXhjZXB0IChPU0Vycm9yLCBSdW50aW1lRXJyb3IsIFZhbHVlRXJyb3IpOgorICAgICAgICByZXR1cm4gTm9uZQorICAgIG1hdGNoID0gY2xhc3NpZnlfc2VjcmV0X3BhdGgoc3RyKHJlc29sdmVkKSwgY3dkPWN3ZCwgaG9tZV9kaXI9aG9tZV9kaXIpCisgICAgcmV0dXJuIG1hdGNoLnBhdGggaWYgbWF0Y2ggaXMgbm90IE5vbmUgZWxzZSBOb25lCisKKworZGVmIF91bndyYXBfZXhlY3V0aW9uX2J1aWx0aW4oCisgICAgZXhlY3V0YWJsZTogc3RyLAorICAgIGFyZ3M6IHR1cGxlW3N0ciwgLi4uXSwKKykgLT4gdHVwbGVbc3RyIHwgTm9uZSwgdHVwbGVbc3RyLCAuLi5dLCBib29sXToKKyAgICAiIiJVbndyYXAgbGl0ZXJhbCBjb21tYW5kL2V4ZWMgcHJlZml4ZXMgb3IgcmVwb3J0IGFtYmlndW91cyBleGVjdXRpb24uIiIiCisKKyAgICBuYW1lID0gUGF0aChleGVjdXRhYmxlIG9yICIiKS5uYW1lLmxvd2VyKCkKKyAgICBpZiBuYW1lID09ICJjb21tYW5kIjoKKyAgICAgICAgaW5kZXggPSAwCisgICAgICAgIGxvb2t1cF9vbmx5ID0gRmFsc2UKKyAgICAgICAgd2hpbGUgaW5kZXggPCBsZW4oYXJncyk6CisgICAgICAgICAgICBhcmcgPSBhcmdzW2luZGV4XQorICAgICAgICAgICAgaWYgYXJnID09ICItLSI6CisgICAgICAgICAgICAgICAgaW5kZXggKz0gMQorICAgICAgICAgICAgICAgIGJyZWFrCisgICAgICAgICAgICBpZiBhcmcuc3RhcnRzd2l0aCgiLSIpOgorICAgICAgICAgICAgICAgIGZsYWdzID0gYXJnWzE6XQorICAgICAgICAgICAgICAgIGlmIG5vdCBmbGFncyBvciBhbnkoZmxhZyBub3QgaW4gInBWdiIgZm9yIGZsYWcgaW4gZmxhZ3MpOgorICAgICAgICAgICAgICAgICAgICByZXR1cm4gTm9uZSwgKCksIFRydWUKKyAgICAgICAgICAgICAgICBsb29rdXBfb25seSA9IGxvb2t1cF9vbmx5IG9yICJ2IiBpbiBmbGFncyBvciAiViIgaW4gZmxhZ3MKKyAgICAgICAgICAgICAgICBpbmRleCArPSAxCisgICAgICAgICAgICAgICAgY29udGludWUKKyAgICAgICAgICAgIGJyZWFrCisgICAgICAgIGlmIGxvb2t1cF9vbmx5OgorICAgICAgICAgICAgcmV0dXJuIE5vbmUsICgpLCBGYWxzZQorICAgICAgICBpZiBpbmRleCA+PSBsZW4oYXJncyk6CisgICAgICAgICAgICByZXR1cm4gTm9uZSwgKCksIEZhbHNlCisgICAgICAgIHJldHVybiBhcmdzW2luZGV4XSwgYXJnc1tpbmRleCArIDEgOl0sIEZhbHNlCisgICAgaWYgbmFtZSA9PSAiZXhlYyI6CisgICAgICAgIGluZGV4ID0gMAorICAgICAgICB3aGlsZSBpbmRleCA8IGxlbihhcmdzKToKKyAgICAgICAgICAgIGFyZyA9IGFyZ3NbaW5kZXhdCisgICAgICAgICAgICBpZiBhcmcgPT0gIi0tIjoKKyAgICAgICAgICAgICAgICBpbmRleCArPSAxCisgICAgICAgICAgICAgICAgYnJlYWsKKyAgICAgICAgICAgIGlmIGFyZyA9PSAiLWEiOgorICAgICAgICAgICAgICAgIGlmIGluZGV4ICsgMSA+PSBsZW4oYXJncyk6CisgICAgICAgICAgICAgICAgICAgIHJldHVybiBOb25lLCAoKSwgVHJ1ZQorICAgICAgICAgICAgICAgIGluZGV4ICs9IDIKKyAgICAgICAgICAgICAgICBjb250aW51ZQorICAgICAgICAgICAgaWYgYXJnIGluIHsiLWMiLCAiLWwifToKKyAgICAgICAgICAgICAgICBpbmRleCArPSAxCisgICAgICAgICAgICAgICAgY29udGludWUKKyAgICAgICAgICAgIGlmIGFyZy5zdGFydHN3aXRoKCItIik6CisgICAgICAgICAgICAgICAgcmV0dXJuIE5vbmUsICgpLCBUcnVlCisgICAgICAgICAgICBicmVhaworICAgICAgICBpZiBpbmRleCA+PSBsZW4oYXJncyk6CisgICAgICAgICAgICByZXR1cm4gTm9uZSwgKCksIEZhbHNlCisgICAgICAgIHJldHVybiBhcmdzW2luZGV4XSwgYXJnc1tpbmRleCArIDEgOl0sIEZhbHNlCisgICAgcmV0dXJuIGV4ZWN1dGFibGUsIGFyZ3MsIEZhbHNlCisKKworZGVmIF9kaXJlY3Rfc2VjcmV0X3JlYWRfcGF0aHNfZnJvbV90b2tlbnMoCisgICAgdG9rZW5zOiB0dXBsZVtzdHIsIC4uLl0sCisgICAgKiwKKyAgICBjd2Q6IFBhdGggfCBOb25lLAorICAgIGhvbWVfZGly
