--- .github/pr2870-finalize/002.b64
@@ -1,1 +1,0 @@
-OiBQYXRoIHwgTm9uZSwKKykgLT4gdHVwbGVbc3RyLCAuLi5dOgorICAgICIiIlJlY292ZXIgZGlyZWN0IHJlYWRlciBvcGVyYW5kcyB3aGVuIHRoZSBmdWxsIGNvbW1hbmQgZXhjZWVkcyBwYXJzZXIgYm91bmRzLiIiIgorCisgICAgZnJvbSAuc2VjcmV0X2ZpbGVfcmVxdWVzdF9zZXJ2aWNlcy5sb2NhbF9yZWFkX29wZXJhbmRzIGltcG9ydCBfc2hlbGxfc2VnbWVudF9maWxlX29wZXJhbmRfdG9rZW5zCisKKyAgICBpZiBub3QgdG9rZW5zOgorICAgICAgICByZXR1cm4gKCkKKyAgICBleGVjdXRhYmxlLCBhcmdzLCBhbWJpZ3VvdXMgPSBfdW53cmFwX2V4ZWN1dGlvbl9idWlsdGluKHRva2Vuc1swXSwgdG9rZW5zWzE6XSkKKyAgICBpZiBhbWJpZ3VvdXMgb3IgZXhlY3V0YWJsZSBpcyBOb25lOgorICAgICAgICByZXR1cm4gKCkKKyAgICBuYW1lID0gIi4iIGlmIGV4ZWN1dGFibGUgPT0gIi4iIGVsc2UgUGF0aChleGVjdXRhYmxlKS5uYW1lLmxvd2VyKCkKKyAgICBhcmdzX2xpc3QgPSBsaXN0KGFyZ3MpCisgICAgY2FuZGlkYXRlcyA9IGxpc3QoX3NoZWxsX3NlZ21lbnRfZmlsZV9vcGVyYW5kX3Rva2VucyhbbmFtZSwgKmFyZ3NfbGlzdF0pKQorICAgIGlmIG5hbWUgaW4gX09USEVSX1JFQURFUlM6CisgICAgICAgIGNhbmRpZGF0ZXMuZXh0ZW5kKGFyZyBmb3IgYXJnIGluIGFyZ3NfbGlzdCBpZiBub3QgYXJnLnN0YXJ0c3dpdGgoIi0iKSkKKyAgICBpZiBuYW1lIGluIHsic291cmNlIiwgIi4ifToKKyAgICAgICAgY2FuZGlkYXRlcy5leHRlbmQoYXJnc19saXN0WzoxXSkKKyAgICBmb3IgaW5kZXgsIHRva2VuIGluIGVudW1lcmF0ZShhcmdzX2xpc3RbOi0xXSk6CisgICAgICAgIGlmIHJlLmZ1bGxtYXRjaChyIlxkKjx8XGQqPD4iLCB0b2tlbik6CisgICAgICAgICAgICBjYW5kaWRhdGVzLmFwcGVuZChhcmdzX2xpc3RbaW5kZXggKyAxXSkKKyAgICBmb3IgdG9rZW4gaW4gYXJnc19saXN0OgorICAgICAgICBtYXRjaCA9IHJlLmZ1bGxtYXRjaChyIlxkKjwoPyE8KSguKykiLCB0b2tlbikKKyAgICAgICAgaWYgbWF0Y2ggaXMgbm90IE5vbmU6CisgICAgICAgICAgICBjYW5kaWRhdGVzLmFwcGVuZChtYXRjaC5ncm91cCgxKSkKKyAgICByZXR1cm4gdHVwbGUoCisgICAgICAgIGRpY3QuZnJvbWtleXMoCisgICAgICAgICAgICBwYXRoIGZvciB2YWx1ZSBpbiBjYW5kaWRhdGVzIGlmIChwYXRoIDo9IF9zZW5zaXRpdmVfcGF0aCh2YWx1ZSwgY3dkPWN3ZCwgaG9tZV9kaXI9aG9tZV9kaXIpKSBpcyBub3QgTm9uZQorICAgICAgICApCisgICAgKQorCisKK2RlZiBkaXJlY3Rfc2VjcmV0X3JlYWRfcGF0aHMoCisgICAgY29tbWFuZDogQ2Fub25pY2FsQ29tbWFuZCwKKyAgICAqLAorICAgIGN3ZDogUGF0aCB8IE5vbmUsCisgICAgaG9tZV9kaXI6IFBhdGggfCBOb25lLAorKSAtPiB0dXBsZVtzdHIsIC4uLl06CisgICAgIiIiRmluZCBkaXJlY3QgY3JlZGVudGlhbCByZWFkcyBmb3IgYSBjb21tYW5kIHdob3NlIGN3ZCBpcyBhbHJlYWR5IHByb3Zlbi4iIiIKKworICAgIGZyb20gLnNlY3JldF9maWxlX3JlcXVlc3Rfc2VydmljZXMubG9jYWxfcmVhZF9vcGVyYW5kcyBpbXBvcnQgX3NoZWxsX3NlZ21lbnRfZmlsZV9vcGVyYW5kX3Rva2VucworCisgICAgY2FuZGlkYXRlczogbGlzdFtzdHJdID0gW10KKyAgICBmb3Igc2VnbWVudCBpbiBjb21tYW5kLnNlZ21lbnRzOgorICAgICAgICBleGVjdXRhYmxlLCBhcmdzLCBhbWJpZ3VvdXMgPSBfdW53cmFwX2V4ZWN1dGlvbl9idWlsdGluKAorICAgICAgICAgICAgc2VnbWVudC5leGVjdXRhYmxlIG9yICIiLAorICAgICAgICAgICAgc2VnbWVudC5hcmd1bWVudHMsCisgICAgICAgICkKKyAgICAgICAgaWYgYW1iaWd1b3VzIG9yIGV4ZWN1dGFibGUgaXMgTm9uZToKKyAgICAgICAgICAgIGNvbnRpbnVlCisgICAgICAgIG5hbWUgPSAiLiIgaWYgZXhlY3V0YWJsZSA9PSAiLiIgZWxzZSBQYXRoKGV4ZWN1dGFibGUpLm5hbWUubG93ZXIoKQorICAgICAgICBhcmdzX2xpc3QgPSBsaXN0KGFyZ3MpCisgICAgICAgIGNhbmRpZGF0ZXMuZXh0ZW5kKF9zaGVsbF9zZWdtZW50X2ZpbGVfb3BlcmFuZF90b2tlbnMoW25hbWUsICphcmdzX2xpc3RdKSkKKyAgICAgICAgaWYgbmFtZSBpbiBfT1RIRVJfUkVBREVSUzoKKyAgICAgICAgICAgIGNhbmRpZGF0ZXMuZXh0ZW5kKGFyZyBmb3IgYXJnIGluIGFyZ3NfbGlzdCBpZiBub3QgYXJnLnN0YXJ0c3dpdGgoIi0iKSkKKyAgICAgICAgaWYgbmFtZSBpbiB7InNvdXJjZSIsICIuIn06CisgICAgICAgICAgICBjYW5kaWRhdGVzLmV4dGVuZChhcmdzX2xpc3RbOjFdKQorICAgICAgICBpZiBuYW1lIGluIHsibm9kZSIsICJidW4iLCAicnVieSIsICJwZXJsIn0gb3IgX3B5dGhvbl9leGVjdXRhYmxlKG5hbWUpOgorICAgICAgICAgICAgZm9yIGluZGV4LCBhcmcgaW4gZW51bWVyYXRlKGFyZ3NfbGlzdFs6LTFdKToKKyAgICAgICAgICAgICAgICBpZiBhcmcgaW4geyItYyIsICItZSIsICItLWV2YWwiLCAiLXAiLCAiLS1wcmludCJ9OgorICAgICAgICAgICAgICAgICAgICBjYW5kaWRhdGVzLmV4dGVuZChfbGl0ZXJhbF9yZWFkX3BhdGhzKGFyZ3NfbGlzdFtpbmRleCArIDFdKSkKKyAgICAgICAgICAgICAgICBlbGlmIGFyZy5zdGFydHN3aXRoKCgiLS1ldmFsPSIsICItLXByaW50PSIpKToKKyAgICAgICAgICAgICAgICAgICAgY2FuZGlkYXRlcy5leHRlbmQoX2xpdGVyYWxfcmVhZF9wYXRocyhhcmcuc3BsaXQoIj0iLCAxKVsxXSkpCisgICAgZm9yIHJlZGlyZWN0IGluIGNvbW1hbmQucmVkaXJlY3RzOgorICAgICAgICBpZiByZWRpcmVjdC5vcGVyYXRvci5sc3RyaXAoIjAxMjM0NTY3ODkiKSBpbiB7IjwiLCAiPD4ifToKKyAgICAgICAgICAgIGNhbmRpZGF0ZXMuYXBwZW5kKHJlZGlyZWN0LnRhcmdldCkKKyAgICByZXR1cm4gdHVwbGUoCisgICAgICAgIGRpY3QuZnJvbWtleXMoCisgICAgICAgICAgICBwYXRoIGZvciB2YWx1ZSBpbiBjYW5kaWRhdGVzIGlmIChwYXRoIDo9IF9zZW5zaXRpdmVfcGF0aCh2YWx1ZSwg
