--- .github/pr2870-finalize/003.b64
@@ -1,1 +1,0 @@
-Y3dkPWN3ZCwgaG9tZV9kaXI9aG9tZV9kaXIpKSBpcyBub3QgTm9uZQorICAgICAgICApCisgICAgKQorCisKK2RlZiBfc2hlbGxfY29tbWFuZF9zdHJpbmcoZXhlY3V0YWJsZTogc3RyLCBhcmdzOiB0dXBsZVtzdHIsIC4uLl0pIC0+IHR1cGxlW3N0ciB8IE5vbmUsIGJvb2xdOgorICAgICIiIlJldHVybiBhIGxpdGVyYWwgc2hlbGwgY29tbWFuZCBzdHJpbmcgYW5kIHdoZXRoZXIgY29tbWFuZC1zdHJpbmcgbW9kZSB3YXMgcmVxdWVzdGVkLiIiIgorCisgICAgbmFtZSA9ICIuIiBpZiBleGVjdXRhYmxlID09ICIuIiBlbHNlIFBhdGgoZXhlY3V0YWJsZSkubmFtZS5sb3dlcigpCisgICAgaWYgbmFtZSBub3QgaW4gX1NIRUxMUzoKKyAgICAgICAgcmV0dXJuIE5vbmUsIEZhbHNlCisgICAgaW5kZXggPSAwCisgICAgd2hpbGUgaW5kZXggPCBsZW4oYXJncyk6CisgICAgICAgIGFyZyA9IGFyZ3NbaW5kZXhdCisgICAgICAgIGlmIGFyZyA9PSAiLS0iOgorICAgICAgICAgICAgcmV0dXJuIE5vbmUsIEZhbHNlCisgICAgICAgIGlmIGFyZyA9PSAiLS1jb21tYW5kIjoKKyAgICAgICAgICAgIHJldHVybiAoYXJnc1tpbmRleCArIDFdIGlmIGluZGV4ICsgMSA8IGxlbihhcmdzKSBlbHNlIE5vbmUpLCBUcnVlCisgICAgICAgIGlmIGFyZy5zdGFydHN3aXRoKCItLWNvbW1hbmQ9Iik6CisgICAgICAgICAgICByZXR1cm4gYXJnLnNwbGl0KCI9IiwgMSlbMV0sIFRydWUKKyAgICAgICAgaWYgYXJnID09ICItcyI6CisgICAgICAgICAgICByZXR1cm4gTm9uZSwgRmFsc2UKKyAgICAgICAgaWYgYXJnLnN0YXJ0c3dpdGgoIi0iKSBhbmQgbm90IGFyZy5zdGFydHN3aXRoKCItLSIpIGFuZCAiYyIgaW4gYXJnWzE6XToKKyAgICAgICAgICAgIGZyb20gLmludGVycHJldGVyX29wdGlvbnMgaW1wb3J0IHNoZWxsX2ludGVycHJldGVyX2NvbW1hbmRfcGF5bG9hZAorCisgICAgICAgICAgICBwYXJzZWQgPSBzaGVsbF9pbnRlcnByZXRlcl9jb21tYW5kX3BheWxvYWQoW2V4ZWN1dGFibGUsICphcmdzXSwgMCkKKyAgICAgICAgICAgIHJldHVybiAocGFyc2VkLnNjcmlwdF90ZXh0IGlmIHBhcnNlZCBpcyBub3QgTm9uZSBlbHNlIE5vbmUpLCBUcnVlCisgICAgICAgIGlmIGFyZyBpbiB7Ii1vIiwgIi1PIiwgIitvIiwgIitPIiwgIi0tcmNmaWxlIiwgIi0taW5pdC1maWxlIn06CisgICAgICAgICAgICBpbmRleCArPSAyCisgICAgICAgICAgICBjb250aW51ZQorICAgICAgICBpZiBub3QgYXJnLnN0YXJ0c3dpdGgoKCItIiwgIisiKSk6CisgICAgICAgICAgICByZXR1cm4gTm9uZSwgRmFsc2UKKyAgICAgICAgaW5kZXggKz0gMQorICAgIHJldHVybiBOb25lLCBGYWxzZQorCisKK2RlZiBfc2NyaXB0X29wZXJhbmQoZXhlY3V0YWJsZTogc3RyLCBhcmdzOiB0dXBsZVtzdHIsIC4uLl0pIC0+IHR1cGxlW3N0ciwgYm9vbF0gfCBOb25lOgorICAgICIiIlJldHVybiBvbmUgbGl0ZXJhbCBsb2NhbCBzY3JpcHQgb3BlcmFuZCB0aGF0IGNhbiBiZSBpbnNwZWN0ZWQgc2FmZWx5LiIiIgorCisgICAgbmFtZSA9ICIuIiBpZiBleGVjdXRhYmxlID09ICIuIiBlbHNlIFBhdGgoZXhlY3V0YWJsZSkubmFtZS5sb3dlcigpCisgICAgaXNfc2hlbGwgPSBuYW1lIGluIF9TSEVMTFMKKyAgICBpc19pbnRlcnByZXRlciA9IG5hbWUgaW4geyJub2RlIiwgInJ1YnkiLCAicGVybCJ9IG9yIF9weXRob25fZXhlY3V0YWJsZShuYW1lKQorICAgIGlmIGlzX3NoZWxsIG9yIGlzX2ludGVycHJldGVyOgorICAgICAgICBpbmRleCA9IDAKKyAgICAgICAgd2hpbGUgaW5kZXggPCBsZW4oYXJncyk6CisgICAgICAgICAgICBhcmcgPSBhcmdzW2luZGV4XQorICAgICAgICAgICAgaWYgKAorICAgICAgICAgICAgICAgIGFyZyBpbiB7Ii1jIiwgIi1lIiwgIi0tZXZhbCIsICItbSIsICItLWNvbW1hbmQifQorICAgICAgICAgICAgICAgIG9yIChpc19zaGVsbCBhbmQgYXJnID09ICItcyIpCisgICAgICAgICAgICAgICAgb3IgKGlzX3NoZWxsIGFuZCBhcmcuc3RhcnRzd2l0aCgiLSIpIGFuZCBub3QgYXJnLnN0YXJ0c3dpdGgoIi0tIikgYW5kICJjIiBpbiBhcmdbMTpdKQorICAgICAgICAgICAgKToKKyAgICAgICAgICAgICAgICByZXR1cm4gTm9uZQorICAgICAgICAgICAgaWYgYXJnID09ICItLSI6CisgICAgICAgICAgICAgICAgaW5kZXggKz0gMQorICAgICAgICAgICAgICAgIGJyZWFrCisgICAgICAgICAgICBpZiBhcmcgaW4geyItbyIsICItTyIsICIrbyIsICIrTyIsICItLXJjZmlsZSIsICItLWluaXQtZmlsZSJ9OgorICAgICAgICAgICAgICAgIGluZGV4ICs9IDIKKyAgICAgICAgICAgICAgICBjb250aW51ZQorICAgICAgICAgICAgaWYgbm90IGFyZy5zdGFydHN3aXRoKCgiLSIsICIrIikpOgorICAgICAgICAgICAgICAgIGJyZWFrCisgICAgICAgICAgICBpbmRleCArPSAxCisgICAgICAgIGlmIGluZGV4IDwgbGVuKGFyZ3MpIGFuZCBhcmdzW2luZGV4XSAhPSAiLSI6CisgICAgICAgICAgICByZXR1cm4gYXJnc1tpbmRleF0sIGlzX3NoZWxsCisgICAgaWYgbmFtZSA9PSAiYnVuIiBhbmQgYXJncyBhbmQgYXJnc1swXS5lbmRzd2l0aChfU0NSSVBUX1NVRkZJWEVTKToKKyAgICAgICAgcmV0dXJuIGFyZ3NbMF0sIEZhbHNlCisgICAgaWYgZXhlY3V0YWJsZS5lbmRzd2l0aChfU0NSSVBUX1NVRkZJWEVTKSBhbmQgKCIvIiBpbiBleGVjdXRhYmxlIG9yIGV4ZWN1dGFibGUuc3RhcnRzd2l0aCgiLiIpKToKKyAgICAgICAgcmV0dXJuIGV4ZWN1dGFibGUsIGV4ZWN1dGFibGUuZW5kc3dpdGgoKCIuc2giLCAiLmJhc2giLCAiLnpzaCIsICIua3NoIiwgIi5maXNoIikpCisgICAgcmV0dXJuIE5vbmUKKworCitkZWYgX2tub3duX3B5dGhvbl9tb2R1bGVfbGF1bmNoKGV4ZWN1dGFibGU6IHN0ciwgYXJnczogdHVwbGVbc3RyLCAuLi5dLCAqLCBjd2Q6IFBhdGggfCBOb25lKSAtPiBib29sOgorICAgICIiIlJldHVybiBUcnVlIGZvciBhIHRydXN0ZWQgdG9vbGluZyBtb2R1bGUgaGFuZGxlZCBieSBkZWRpY2F0ZWQgcG9saWN5LiIiIgorCisgICAg
