--- .github/pr2870-finalize/004.b64
@@ -1,1 +1,0 @@
-bmFtZSA9IFBhdGgoZXhlY3V0YWJsZSBvciAiIikubmFtZS5sb3dlcigpCisgICAgaWYgbm90IF9weXRob25fZXhlY3V0YWJsZShuYW1lKToKKyAgICAgICAgcmV0dXJuIEZhbHNlCisgICAgZnJvbSAuc2VjcmV0X2ZpbGVfcmVxdWVzdF9zZXJ2aWNlcy5jb25zdGFudHNfY29yZSBpbXBvcnQgX1NBRkVfUFlUSE9OX01PRFVMRV9DT01NQU5EUworICAgIGZyb20gLnNlY3JldF9maWxlX3JlcXVlc3Rfc2VydmljZXMuaW50ZXJwcmV0ZXJfb2JzZXJ2ZXJzIGltcG9ydCBfcHl0aG9uX21vZHVsZV9tYXlfYmVfc2hhZG93ZWQKKyAgICBmcm9tIC5zZWNyZXRfZmlsZV9yZXF1ZXN0X3NlcnZpY2VzLnB5dGVzdF9jb25maWdfc2FmZXR5IGltcG9ydCBfcHl0aG9uX21vZHVsZV9yb290X2Zyb21fYXJncworCisgICAgbW9kdWxlX3Jvb3QgPSBfcHl0aG9uX21vZHVsZV9yb290X2Zyb21fYXJncyhsaXN0KGFyZ3MpKQorICAgIHJldHVybiBib29sKG1vZHVsZV9yb290IGluIF9TQUZFX1BZVEhPTl9NT0RVTEVfQ09NTUFORFMgYW5kIG5vdCBfcHl0aG9uX21vZHVsZV9tYXlfYmVfc2hhZG93ZWQobW9kdWxlX3Jvb3QsIGN3ZCkpCisKKworZGVmIF9weXRob25fbW9kdWxlX2xhdW5jaChleGVjdXRhYmxlOiBzdHIsIGFyZ3M6IHR1cGxlW3N0ciwgLi4uXSwgKiwgY3dkOiBQYXRoIHwgTm9uZSkgLT4gYm9vbDoKKyAgICAiIiJSZXR1cm4gVHJ1ZSB3aGVuIFB5dGhvbiBtYXkgaW1wb3J0IG11dGFibGUgd29ya3NwYWNlIGNvZGUgdmlhIGBgLW1gYC4iIiIKKworICAgIG5hbWUgPSBQYXRoKGV4ZWN1dGFibGUgb3IgIiIpLm5hbWUubG93ZXIoKQorICAgIGlmIG5vdCBfcHl0aG9uX2V4ZWN1dGFibGUobmFtZSk6CisgICAgICAgIHJldHVybiBGYWxzZQorICAgIGlmIF9rbm93bl9weXRob25fbW9kdWxlX2xhdW5jaChleGVjdXRhYmxlLCBhcmdzLCBjd2Q9Y3dkKToKKyAgICAgICAgcmV0dXJuIEZhbHNlCisgICAgZm9yIGluZGV4LCBhcmcgaW4gZW51bWVyYXRlKGFyZ3MpOgorICAgICAgICBpZiBhcmcgPT0gIi0tIjoKKyAgICAgICAgICAgIHJldHVybiBGYWxzZQorICAgICAgICBpZiBhcmcgPT0gIi1tIiBvciAoYXJnLnN0YXJ0c3dpdGgoIi1tIikgYW5kIGxlbihhcmcpID4gMik6CisgICAgICAgICAgICByZXR1cm4gVHJ1ZQorICAgICAgICBpZiBhcmcgaW4geyItVyIsICItWCJ9IGFuZCBpbmRleCArIDEgPCBsZW4oYXJncyk6CisgICAgICAgICAgICBjb250aW51ZQorICAgICAgICBpZiBub3QgYXJnLnN0YXJ0c3dpdGgoIi0iKToKKyAgICAgICAgICAgIHJldHVybiBGYWxzZQorICAgIHJldHVybiBGYWxzZQorCisKK2RlZiBfaW50ZXJwcmV0ZXJfc3RkaW5fbGF1bmNoKGV4ZWN1dGFibGU6IHN0ciwgYXJnczogdHVwbGVbc3RyLCAuLi5dKSAtPiBib29sOgorICAgICIiIlJldHVybiBUcnVlIHdoZW4gYW4gaW50ZXJwcmV0ZXIgZXhlY3V0ZXMgc291cmNlIHN1cHBsaWVkIG9uIHN0ZGluLiIiIgorCisgICAgbmFtZSA9IFBhdGgoZXhlY3V0YWJsZSBvciAiIikubmFtZS5sb3dlcigpCisgICAgaWYgbmFtZSBub3QgaW4geyJub2RlIiwgImJ1biIsICJydWJ5IiwgInBlcmwifSBhbmQgbm90IF9weXRob25fZXhlY3V0YWJsZShuYW1lKToKKyAgICAgICAgcmV0dXJuIEZhbHNlCisgICAgaW5saW5lX2ZsYWdzID0geyItYyIsICItZSIsICItLWV2YWwiLCAiLXAiLCAiLS1wcmludCJ9CisgICAgaWYgYW55KGFyZyBpbiBpbmxpbmVfZmxhZ3Mgb3IgYXJnLnN0YXJ0c3dpdGgoKCItLWV2YWw9IiwgIi0tcHJpbnQ9IikpIGZvciBhcmcgaW4gYXJncyk6CisgICAgICAgIHJldHVybiBGYWxzZQorICAgIHJldHVybiAiLSIgaW4gYXJncworCisKK2RlZiBfaW50ZXJwcmV0ZXJfaW5saW5lX2xhdW5jaChleGVjdXRhYmxlOiBzdHIsIGFyZ3M6IHR1cGxlW3N0ciwgLi4uXSkgLT4gYm9vbDoKKyAgICAiIiJSZXR1cm4gVHJ1ZSBmb3IgbGl0ZXJhbCBpbmxpbmUtY29kZSBtb2RlcyB3aG9zZSBmaWxlIHJlYWRzIHdlcmUgYWxyZWFkeSBzY2FubmVkLiIiIgorCisgICAgbmFtZSA9IFBhdGgoZXhlY3V0YWJsZSBvciAiIikubmFtZS5sb3dlcigpCisgICAgaWYgbmFtZSBub3QgaW4geyJub2RlIiwgImJ1biIsICJydWJ5IiwgInBlcmwifSBhbmQgbm90IF9weXRob25fZXhlY3V0YWJsZShuYW1lKToKKyAgICAgICAgcmV0dXJuIEZhbHNlCisgICAgZm9yIGFyZyBpbiBhcmdzOgorICAgICAgICBpZiBhcmcgPT0gIi0tIjoKKyAgICAgICAgICAgIHJldHVybiBGYWxzZQorICAgICAgICBpZiBhcmcgaW4geyItYyIsICItZSIsICItLWV2YWwiLCAiLXAiLCAiLS1wcmludCJ9IG9yIGFyZy5zdGFydHN3aXRoKCgiLS1ldmFsPSIsICItLXByaW50PSIpKToKKyAgICAgICAgICAgIHJldHVybiBUcnVlCisgICAgICAgIGlmIG5vdCBhcmcuc3RhcnRzd2l0aCgiLSIpOgorICAgICAgICAgICAgcmV0dXJuIEZhbHNlCisgICAgcmV0dXJuIEZhbHNlCisKKworZGVmIF9wYXRoX3F1YWxpZmllZChleGVjdXRhYmxlOiBzdHIpIC0+IGJvb2w6CisgICAgcmV0dXJuIGJvb2woZXhlY3V0YWJsZSkgYW5kICgiLyIgaW4gZXhlY3V0YWJsZSBvciAiXFwiIGluIGV4ZWN1dGFibGUgb3IgZXhlY3V0YWJsZS5zdGFydHN3aXRoKCIuIikpCisKKworZGVmIF9sb2NhbF9leGVjdXRhYmxlX29wZXJhbmQoCisgICAgZXhlY3V0YWJsZTogc3RyLAorICAgICosCisgICAgY3dkOiBQYXRoIHwgTm9uZSwKKyAgICBob21lX2RpcjogUGF0aCB8IE5vbmUsCisgICAgcm9vdHM6IHR1cGxlW1BhdGgsIC4uLl0sCispIC0+IHN0ciB8IE5vbmU6CisgICAgIiIiUmV0dXJuIGEgcGF0aC1xdWFsaWZpZWQgZXhlY3V0YWJsZSBvbmx5IHdoZW4gaXQgaXMgaW5zaWRlIGEgZ3VhcmRlZCBsb2NhbCByb290LiIiIgorCisgICAgaWYgbm90IF9wYXRoX3F1YWxpZmllZChleGVjdXRhYmxlKToKKyAgICAgICAgcmV0dXJuIE5vbmUKKyAgICBsZXhpY2FsID0gUGF0aChub3JtYWxpemVfcGF0aChleHBhbmRfaG9tZShleGVjdXRhYmxlLCBob21lX2RpciksIGN3ZCkpCisgICAgaWYgbm90IGxl
