--- .github/pr2870-finalize/005.b64
@@ -1,1 +1,0 @@
-eGljYWwuaXNfYWJzb2x1dGUoKSBvciBub3QgYW55KGxleGljYWwuaXNfcmVsYXRpdmVfdG8ocm9vdCkgZm9yIHJvb3QgaW4gcm9vdHMpOgorICAgICAgICByZXR1cm4gTm9uZQorICAgIHJldHVybiBleGVjdXRhYmxlCisKKworZGVmIF9wYXJzZV9leGVjdXRpb25fc2VnbWVudCgKKyAgICBleGVjdXRpb246IFNoZWxsRXhlY3V0aW9uU2VnbWVudCwKKyAgICAqLAorICAgIHJhd19tb2RlbDogQ2Fub25pY2FsQ29tbWFuZCwKKyAgICByYXdfc2VnbWVudDogQ29tbWFuZFNlZ21lbnQgfCBOb25lLAorKSAtPiBDYW5vbmljYWxDb21tYW5kIHwgTm9uZToKKyAgICAiIiJLZWVwIHNvdXJjZSBzeW50YXggYW5kIHJlcXVpcmUgYWdyZWVtZW50IHdpdGggdGhlIGN3ZCBtb2RlbC4KKworICAgIFJlLXF1b3RpbmcgZGVjb2RlZCB0b2tlbnMgY2hhbmdlcyBpbnB1dCByZWRpcmVjdGlvbnMgaW50byBvcmRpbmFyeSBhcmd2CisgICAgYW5kIGNhbiBlcmFzZSBjb21tYW5kL2V4ZWMgYW1iaWd1aXR5LiBUaGUgcmF3IGNvbW1hbmQgcGFyc2VyIGRlbGliZXJhdGVseQorICAgIHNraXBzIHRyYW5zcGFyZW50LXdyYXBwZXIgbm9ybWFsaXphdGlvbiBmb3IgdGhpcyBpbnNwZWN0aW9uIHBhdGguCisgICAgIiIiCisKKyAgICBpZiBub3QgZXhlY3V0aW9uLmNvbXBsZXRlIG9yIGV4ZWN1dGlvbi5lZmZlY3RpdmVfY3dkIGlzIE5vbmUgb3IgcmF3X3NlZ21lbnQgaXMgTm9uZToKKyAgICAgICAgcmV0dXJuIE5vbmUKKyAgICB0cnk6CisgICAgICAgIGlmIHNwbGl0X3NoZWxsX3Rva2VucyhyYXdfc2VnbWVudC50ZXh0KSAhPSBleGVjdXRpb24udG9rZW5zOgorICAgICAgICAgICAgcmV0dXJuIE5vbmUKKyAgICBleGNlcHQgVmFsdWVFcnJvcjoKKyAgICAgICAgcmV0dXJuIE5vbmUKKyAgICByZXR1cm4gcmVwbGFjZSgKKyAgICAgICAgcmF3X21vZGVsLAorICAgICAgICBzZWdtZW50cz0ocmF3X3NlZ21lbnQsKSwKKyAgICAgICAgcmVkaXJlY3RzPXR1cGxlKAorICAgICAgICAgICAgcmVkaXJlY3QKKyAgICAgICAgICAgIGZvciByZWRpcmVjdCBpbiByYXdfbW9kZWwucmVkaXJlY3RzCisgICAgICAgICAgICBpZiByYXdfc2VnbWVudC5zdGFydCA8PSByZWRpcmVjdC5zdGFydCBhbmQgcmVkaXJlY3QuZW5kIDw9IHJhd19zZWdtZW50LmVuZAorICAgICAgICApLAorICAgICAgICBlbWJlZGRlZF9jb21tYW5kcz0oKSwKKyAgICApCisKKworZGVmIF9zZWdtZW50X21heV90b3VjaF9sb2NhbF9kYXRhKGV4ZWN1dGlvbjogU2hlbGxFeGVjdXRpb25TZWdtZW50KSAtPiBib29sOgorICAgICIiIktlZXAgdW5rbm93biBmaWxlL2NvZGUgYWNjZXNzIGNsb3NlZCB3aXRob3V0IHRyZWF0aW5nIHN0ZG91dCBhcyBhIGZpbGUuIiIiCisKKyAgICBmcm9tIC5zZWNyZXRfZmlsZV9yZXF1ZXN0X3NlcnZpY2VzLmxvY2FsX3JlYWRfb3BlcmFuZHMgaW1wb3J0IF9zaGVsbF9zZWdtZW50X2ZpbGVfb3BlcmFuZF90b2tlbnMKKworICAgIGlmIG5vdCBleGVjdXRpb24udG9rZW5zOgorICAgICAgICByZXR1cm4gRmFsc2UKKyAgICB0b2tlbiA9IGV4ZWN1dGlvbi50b2tlbnNbMF0KKyAgICBleGVjdXRhYmxlID0gIi4iIGlmIHRva2VuID09ICIuIiBlbHNlIFBhdGgodG9rZW4pLm5hbWUubG93ZXIoKQorICAgIGFyZ3MgPSBsaXN0KGV4ZWN1dGlvbi50b2tlbnNbMTpdKQorICAgIGhhc19pbnB1dF9yZWRpcmVjdCA9IGFueShyZS5tYXRjaChyIl5cZCo8KD8hPCkiLCBpdGVtKSBmb3IgaXRlbSBpbiBleGVjdXRpb24udG9rZW5zKQorICAgIGlmIGhhc19pbnB1dF9yZWRpcmVjdCBvciBfc2hlbGxfc2VnbWVudF9maWxlX29wZXJhbmRfdG9rZW5zKFtleGVjdXRhYmxlLCAqYXJnc10pOgorICAgICAgICByZXR1cm4gVHJ1ZQorICAgIGlmIGV4ZWN1dGFibGUgaW4gX09USEVSX1JFQURFUlM6CisgICAgICAgIHJldHVybiBhbnkobm90IGFyZy5zdGFydHN3aXRoKCItIikgZm9yIGFyZyBpbiBhcmdzKQorICAgIHJldHVybiAoCisgICAgICAgIGV4ZWN1dGFibGUgaW4geypfU0hFTExTLCAiY29tbWFuZCIsICJleGVjIiwgIm5vZGUiLCAiYnVuIiwgInJ1YnkiLCAicGVybCJ9CisgICAgICAgIG9yIF9weXRob25fZXhlY3V0YWJsZShleGVjdXRhYmxlKQorICAgICAgICBvciBfcGF0aF9xdWFsaWZpZWQodG9rZW4pCisgICAgKQorCisKK2RlZiBfZmxvd19vcGVyYXRvcl9iZWZvcmUoZXhlY3V0aW9uOiBTaGVsbEV4ZWN1dGlvblNlZ21lbnQpIC0+IHN0ciB8IE5vbmU6CisgICAgcmV0dXJuIG5leHQoKHRva2VuIGZvciB0b2tlbiBpbiByZXZlcnNlZChleGVjdXRpb24uY29udHJvbF9iZWZvcmUpIGlmIHRva2VuIGluIF9GTE9XX09QRVJBVE9SUyksIE5vbmUpCisKKworZGVmIF9mYWlsZWRfY2Rfc2hvcnRfY2lyY3VpdF9zdGF0ZSgKKyAgICBleGVjdXRpb246IFNoZWxsRXhlY3V0aW9uU2VnbWVudCwKKyAgICAqLAorICAgIGFjdGl2ZTogYm9vbCwKKykgLT4gdHVwbGVbYm9vbCwgYm9vbF06CisgICAgIiIiUmV0dXJuIHRoZSBmYWlsZWQtY2Qgc3RhdGUgYW5kIHdoZXRoZXIgdGhpcyBzZWdtZW50IGlzIHByb3ZhYmx5IHVucmVhY2hhYmxlLiIiIgorCisgICAgb3BlcmF0b3IgPSBfZmxvd19vcGVyYXRvcl9iZWZvcmUoZXhlY3V0aW9uKQorICAgIGlmIG9wZXJhdG9yIGluIHsifHwiLCAiOyIsICImIn06CisgICAgICAgIGFjdGl2ZSA9IEZhbHNlCisgICAgaWYgZXhlY3V0aW9uLmRpcmVjdG9yeV9vcGVyYXRpb24gaXMgbm90IE5vbmUgYW5kIGV4ZWN1dGlvbi5yZWFzb25fY29kZSBpbiBfU0hPUlRfQ0lSQ1VJVElOR19DRF9GQUlMVVJFUzoKKyAgICAgICAgcmV0dXJuIFRydWUsIEZhbHNlCisgICAgcmV0dXJuIGFjdGl2ZSwgYm9vbChhY3RpdmUgYW5kIG9wZXJhdG9yIGluIHsiJiYiLCAifCJ9KQpkaWZmIC0tZ2l0IGEvc3JjL2NvZGV4X3BsdWdpbl9zY2FubmVyL2d1YXJkL3J1bnRpbWUvc2hlbGxfc2VjcmV0X3JlYWRzLnB5IGIvc3JjL2NvZGV4X3BsdWdpbl9zY2FubmVyL2d1YXJkL3J1bnRpbWUvc2hlbGxfc2VjcmV0X3JlYWRzLnB5CmluZGV4IDky
