--- .github/pr2870-finalize/007.b64
@@ -1,1 +1,0 @@
-dGlvbl9jb250ZXh0CiAKIAogQGRhdGFjbGFzcyhmcm96ZW49VHJ1ZSwgc2xvdHM9VHJ1ZSkKQEAgLTg5LDQwNiArNzAsNiBAQCBjbGFzcyBTaGVsbFJlYWRBc3Nlc3NtZW50OgogICAgICAgICByZXR1cm4gaGFzaGxpYi5zaGEyNTYoanNvbi5kdW1wcyhwYXlsb2FkLCBzZXBhcmF0b3JzPSgiLCIsICI6IikpLmVuY29kZSgpKS5oZXhkaWdlc3QoKQogCiAKLWRlZiBfcHl0aG9uX2V4ZWN1dGFibGUobmFtZTogc3RyKSAtPiBib29sOgotICAgIHJldHVybiBfUFlUSE9OX0VYRUNVVEFCTEUuZnVsbG1hdGNoKG5hbWUpIGlzIG5vdCBOb25lCi0KLQotZGVmIF9jb21tYW5kX21heV9uZWVkX3JlYWRfYXNzZXNzbWVudChjb21tYW5kX3RleHQ6IHN0cikgLT4gYm9vbDoKLSAgICAiIiJTa2lwIGZpbGVzeXN0ZW0gbW9kZWxpbmcgd2hlbiBzeW50YXggY2Fubm90IHJlYWQgZmlsZXMgb3IgbGF1bmNoIGxvY2FsIGNvZGUuIiIiCi0KLSAgICB0cnk6Ci0gICAgICAgIHRva2VucyA9IHNwbGl0X3NoZWxsX3Rva2Vucyhjb21tYW5kX3RleHQpCi0gICAgZXhjZXB0IFZhbHVlRXJyb3I6Ci0gICAgICAgIHJldHVybiBUcnVlCi0gICAgaW50ZXJlc3RpbmcgPSB7Ci0gICAgICAgICpfU0hFTExTLAotICAgICAgICAqX09USEVSX1JFQURFUlMsCi0gICAgICAgICJjYXQiLAotICAgICAgICAiaGVhZCIsCi0gICAgICAgICJ0YWlsIiwKLSAgICAgICAgInNlZCIsCi0gICAgICAgICJncmVwIiwKLSAgICAgICAgImVncmVwIiwKLSAgICAgICAgImZncmVwIiwKLSAgICAgICAgInJnIiwKLSAgICAgICAgInJlYWQiLAotICAgICAgICAiY29tbWFuZCIsCi0gICAgICAgICJleGVjIiwKLSAgICAgICAgIm5vZGUiLAotICAgICAgICAiYnVuIiwKLSAgICAgICAgInJ1YnkiLAotICAgICAgICAicGVybCIsCi0gICAgfQotICAgIGZvciB0b2tlbiBpbiB0b2tlbnM6Ci0gICAgICAgIGlmICI8IiBpbiB0b2tlbjoKLSAgICAgICAgICAgIHJldHVybiBUcnVlCi0gICAgICAgIG5vcm1hbGl6ZWQgPSB0b2tlbi5zdHJpcCgpCi0gICAgICAgIGlmIG5vdCBub3JtYWxpemVkOgotICAgICAgICAgICAgY29udGludWUKLSAgICAgICAgbmFtZSA9ICIuIiBpZiBub3JtYWxpemVkID09ICIuIiBlbHNlIFBhdGgobm9ybWFsaXplZCkubmFtZS5sb3dlcigpCi0gICAgICAgIGlmIG5hbWUgaW4gaW50ZXJlc3Rpbmcgb3IgX3B5dGhvbl9leGVjdXRhYmxlKG5hbWUpOgotICAgICAgICAgICAgcmV0dXJuIFRydWUKLSAgICAgICAgaWYgX3BhdGhfcXVhbGlmaWVkKG5vcm1hbGl6ZWQpOgotICAgICAgICAgICAgcmV0dXJuIFRydWUKLSAgICByZXR1cm4gRmFsc2UKLQotCi1kZWYgX3NlbnNpdGl2ZV9wYXRoKHZhbHVlOiBzdHIsICosIGN3ZDogUGF0aCB8IE5vbmUsIGhvbWVfZGlyOiBQYXRoIHwgTm9uZSkgLT4gc3RyIHwgTm9uZToKLSAgICAiIiJDbGFzc2lmeSBhIGxpdGVyYWwgb3IgbWV0YWRhdGEtcmVzb2x2ZWQgYWxpYXMgd2l0aG91dCBvcGVuaW5nIHNlY3JldHMuIiIiCi0KLSAgICBtYXRjaCA9IGNsYXNzaWZ5X3NlY3JldF9wYXRoKHZhbHVlLCBjd2Q9Y3dkLCBob21lX2Rpcj1ob21lX2RpcikKLSAgICBpZiBtYXRjaCBpcyBub3QgTm9uZToKLSAgICAgICAgcmV0dXJuIG1hdGNoLnBhdGgKLSAgICBsZXhpY2FsID0gUGF0aChub3JtYWxpemVfcGF0aChleHBhbmRfaG9tZSh2YWx1ZSwgaG9tZV9kaXIpLCBjd2QpKQotICAgIHJvb3RzID0gdHVwbGUocm9vdCBmb3Igcm9vdCBpbiAoY3dkLCBob21lX2RpcikgaWYgcm9vdCBpcyBub3QgTm9uZSkKLSAgICBpZiBub3QgbGV4aWNhbC5pc19hYnNvbHV0ZSgpIG9yIG5vdCBhbnkobGV4aWNhbC5pc19yZWxhdGl2ZV90byhyb290KSBmb3Igcm9vdCBpbiByb290cyk6Ci0gICAgICAgIHJldHVybiBOb25lCi0gICAgdHJ5OgotICAgICAgICByZXNvbHZlZCA9IGxleGljYWwucmVzb2x2ZShzdHJpY3Q9VHJ1ZSkKLSAgICBleGNlcHQgKE9TRXJyb3IsIFJ1bnRpbWVFcnJvciwgVmFsdWVFcnJvcik6Ci0gICAgICAgIHJldHVybiBOb25lCi0gICAgbWF0Y2ggPSBjbGFzc2lmeV9zZWNyZXRfcGF0aChzdHIocmVzb2x2ZWQpLCBjd2Q9Y3dkLCBob21lX2Rpcj1ob21lX2RpcikKLSAgICByZXR1cm4gbWF0Y2gucGF0aCBpZiBtYXRjaCBpcyBub3QgTm9uZSBlbHNlIE5vbmUKLQotCi1kZWYgX3Vud3JhcF9leGVjdXRpb25fYnVpbHRpbigKLSAgICBleGVjdXRhYmxlOiBzdHIsCi0gICAgYXJnczogdHVwbGVbc3RyLCAuLi5dLAotKSAtPiB0dXBsZVtzdHIgfCBOb25lLCB0dXBsZVtzdHIsIC4uLl0sIGJvb2xdOgotICAgICIiIlVud3JhcCBsaXRlcmFsIGNvbW1hbmQvZXhlYyBwcmVmaXhlcyBvciByZXBvcnQgYW1iaWd1b3VzIGV4ZWN1dGlvbi4iIiIKLQotICAgIG5hbWUgPSBQYXRoKGV4ZWN1dGFibGUgb3IgIiIpLm5hbWUubG93ZXIoKQotICAgIGlmIG5hbWUgPT0gImNvbW1hbmQiOgotICAgICAgICBpbmRleCA9IDAKLSAgICAgICAgbG9va3VwX29ubHkgPSBGYWxzZQotICAgICAgICB3aGlsZSBpbmRleCA8IGxlbihhcmdzKToKLSAgICAgICAgICAgIGFyZyA9IGFyZ3NbaW5kZXhdCi0gICAgICAgICAgICBpZiBhcmcgPT0gIi0tIjoKLSAgICAgICAgICAgICAgICBpbmRleCArPSAxCi0gICAgICAgICAgICAgICAgYnJlYWsKLSAgICAgICAgICAgIGlmIGFyZy5zdGFydHN3aXRoKCItIik6Ci0gICAgICAgICAgICAgICAgZmxhZ3MgPSBhcmdbMTpdCi0gICAgICAgICAgICAgICAgaWYgbm90IGZsYWdzIG9yIGFueShmbGFnIG5vdCBpbiAicFZ2IiBmb3IgZmxhZyBpbiBmbGFncyk6Ci0gICAgICAgICAgICAgICAgICAgIHJldHVybiBOb25lLCAoKSwgVHJ1ZQotICAgICAgICAgICAgICAgIGxvb2t1cF9vbmx5ID0gbG9va3VwX29ubHkgb3IgInYiIGluIGZsYWdzIG9yICJWIiBpbiBmbGFncwotICAgICAgICAgICAgICAgIGluZGV4ICs9IDEKLSAgICAgICAgICAgICAgICBjb250aW51ZQotICAgICAgICAgICAgYnJlYWsKLSAgICAgICAgaWYgbG9va3VwX29ubHk6Ci0gICAgICAgICAgICByZXR1cm4gTm9uZSwg
