--- .github/pr2870-finalize/008.b64
@@ -1,1 +1,0 @@
-KCksIEZhbHNlCi0gICAgICAgIGlmIGluZGV4ID49IGxlbihhcmdzKToKLSAgICAgICAgICAgIHJldHVybiBOb25lLCAoKSwgRmFsc2UKLSAgICAgICAgcmV0dXJuIGFyZ3NbaW5kZXhdLCBhcmdzW2luZGV4ICsgMSA6XSwgRmFsc2UKLSAgICBpZiBuYW1lID09ICJleGVjIjoKLSAgICAgICAgaW5kZXggPSAwCi0gICAgICAgIHdoaWxlIGluZGV4IDwgbGVuKGFyZ3MpOgotICAgICAgICAgICAgYXJnID0gYXJnc1tpbmRleF0KLSAgICAgICAgICAgIGlmIGFyZyA9PSAiLS0iOgotICAgICAgICAgICAgICAgIGluZGV4ICs9IDEKLSAgICAgICAgICAgICAgICBicmVhawotICAgICAgICAgICAgaWYgYXJnID09ICItYSI6Ci0gICAgICAgICAgICAgICAgaWYgaW5kZXggKyAxID49IGxlbihhcmdzKToKLSAgICAgICAgICAgICAgICAgICAgcmV0dXJuIE5vbmUsICgpLCBUcnVlCi0gICAgICAgICAgICAgICAgaW5kZXggKz0gMgotICAgICAgICAgICAgICAgIGNvbnRpbnVlCi0gICAgICAgICAgICBpZiBhcmcgaW4geyItYyIsICItbCJ9OgotICAgICAgICAgICAgICAgIGluZGV4ICs9IDEKLSAgICAgICAgICAgICAgICBjb250aW51ZQotICAgICAgICAgICAgaWYgYXJnLnN0YXJ0c3dpdGgoIi0iKToKLSAgICAgICAgICAgICAgICByZXR1cm4gTm9uZSwgKCksIFRydWUKLSAgICAgICAgICAgIGJyZWFrCi0gICAgICAgIGlmIGluZGV4ID49IGxlbihhcmdzKToKLSAgICAgICAgICAgIHJldHVybiBOb25lLCAoKSwgRmFsc2UKLSAgICAgICAgcmV0dXJuIGFyZ3NbaW5kZXhdLCBhcmdzW2luZGV4ICsgMSA6XSwgRmFsc2UKLSAgICByZXR1cm4gZXhlY3V0YWJsZSwgYXJncywgRmFsc2UKLQotCi1kZWYgX2RpcmVjdF9zZWNyZXRfcmVhZF9wYXRoc19mcm9tX3Rva2VucygKLSAgICB0b2tlbnM6IHR1cGxlW3N0ciwgLi4uXSwKLSAgICAqLAotICAgIGN3ZDogUGF0aCB8IE5vbmUsCi0gICAgaG9tZV9kaXI6IFBhdGggfCBOb25lLAotKSAtPiB0dXBsZVtzdHIsIC4uLl06Ci0gICAgIiIiUmVjb3ZlciBkaXJlY3QgcmVhZGVyIG9wZXJhbmRzIHdoZW4gdGhlIGZ1bGwgY29tbWFuZCBleGNlZWRzIHBhcnNlciBib3VuZHMuIiIiCi0KLSAgICBmcm9tIC5zZWNyZXRfZmlsZV9yZXF1ZXN0X3NlcnZpY2VzLmxvY2FsX3JlYWRfb3BlcmFuZHMgaW1wb3J0IF9zaGVsbF9zZWdtZW50X2ZpbGVfb3BlcmFuZF90b2tlbnMKLQotICAgIGlmIG5vdCB0b2tlbnM6Ci0gICAgICAgIHJldHVybiAoKQotICAgIGV4ZWN1dGFibGUsIGFyZ3MsIGFtYmlndW91cyA9IF91bndyYXBfZXhlY3V0aW9uX2J1aWx0aW4odG9rZW5zWzBdLCB0b2tlbnNbMTpdKQotICAgIGlmIGFtYmlndW91cyBvciBleGVjdXRhYmxlIGlzIE5vbmU6Ci0gICAgICAgIHJldHVybiAoKQotICAgIG5hbWUgPSAiLiIgaWYgZXhlY3V0YWJsZSA9PSAiLiIgZWxzZSBQYXRoKGV4ZWN1dGFibGUpLm5hbWUubG93ZXIoKQotICAgIGFyZ3NfbGlzdCA9IGxpc3QoYXJncykKLSAgICBjYW5kaWRhdGVzID0gbGlzdChfc2hlbGxfc2VnbWVudF9maWxlX29wZXJhbmRfdG9rZW5zKFtuYW1lLCAqYXJnc19saXN0XSkpCi0gICAgaWYgbmFtZSBpbiBfT1RIRVJfUkVBREVSUzoKLSAgICAgICAgY2FuZGlkYXRlcy5leHRlbmQoYXJnIGZvciBhcmcgaW4gYXJnc19saXN0IGlmIG5vdCBhcmcuc3RhcnRzd2l0aCgiLSIpKQotICAgIGlmIG5hbWUgaW4geyJzb3VyY2UiLCAiLiJ9OgotICAgICAgICBjYW5kaWRhdGVzLmV4dGVuZChhcmdzX2xpc3RbOjFdKQotICAgIGZvciBpbmRleCwgdG9rZW4gaW4gZW51bWVyYXRlKGFyZ3NfbGlzdFs6LTFdKToKLSAgICAgICAgaWYgcmUuZnVsbG1hdGNoKHIiXGQqPHxcZCo8PiIsIHRva2VuKToKLSAgICAgICAgICAgIGNhbmRpZGF0ZXMuYXBwZW5kKGFyZ3NfbGlzdFtpbmRleCArIDFdKQotICAgIGZvciB0b2tlbiBpbiBhcmdzX2xpc3Q6Ci0gICAgICAgIG1hdGNoID0gcmUuZnVsbG1hdGNoKHIiXGQqPCg/ITwpKC4rKSIsIHRva2VuKQotICAgICAgICBpZiBtYXRjaCBpcyBub3QgTm9uZToKLSAgICAgICAgICAgIGNhbmRpZGF0ZXMuYXBwZW5kKG1hdGNoLmdyb3VwKDEpKQotICAgIHJldHVybiB0dXBsZSgKLSAgICAgICAgZGljdC5mcm9ta2V5cygKLSAgICAgICAgICAgIHBhdGggZm9yIHZhbHVlIGluIGNhbmRpZGF0ZXMgaWYgKHBhdGggOj0gX3NlbnNpdGl2ZV9wYXRoKHZhbHVlLCBjd2Q9Y3dkLCBob21lX2Rpcj1ob21lX2RpcikpIGlzIG5vdCBOb25lCi0gICAgICAgICkKLSAgICApCi0KLQotZGVmIGRpcmVjdF9zZWNyZXRfcmVhZF9wYXRocygKLSAgICBjb21tYW5kOiBDYW5vbmljYWxDb21tYW5kLAotICAgICosCi0gICAgY3dkOiBQYXRoIHwgTm9uZSwKLSAgICBob21lX2RpcjogUGF0aCB8IE5vbmUsCi0pIC0+IHR1cGxlW3N0ciwgLi4uXToKLSAgICAiIiJGaW5kIGRpcmVjdCBjcmVkZW50aWFsIHJlYWRzIGZvciBhIGNvbW1hbmQgd2hvc2UgY3dkIGlzIGFscmVhZHkgcHJvdmVuLiIiIgotCi0gICAgZnJvbSAuc2VjcmV0X2ZpbGVfcmVxdWVzdF9zZXJ2aWNlcy5sb2NhbF9yZWFkX29wZXJhbmRzIGltcG9ydCBfc2hlbGxfc2VnbWVudF9maWxlX29wZXJhbmRfdG9rZW5zCi0KLSAgICBjYW5kaWRhdGVzOiBsaXN0W3N0cl0gPSBbXQotICAgIGZvciBzZWdtZW50IGluIGNvbW1hbmQuc2VnbWVudHM6Ci0gICAgICAgIGV4ZWN1dGFibGUsIGFyZ3MsIGFtYmlndW91cyA9IF91bndyYXBfZXhlY3V0aW9uX2J1aWx0aW4oCi0gICAgICAgICAgICBzZWdtZW50LmV4ZWN1dGFibGUgb3IgIiIsCi0gICAgICAgICAgICBzZWdtZW50LmFyZ3VtZW50cywKLSAgICAgICAgKQotICAgICAgICBpZiBhbWJpZ3VvdXMgb3IgZXhlY3V0YWJsZSBpcyBOb25lOgotICAgICAgICAgICAgY29udGludWUKLSAgICAgICAgbmFtZSA9ICIuIiBpZiBleGVjdXRhYmxlID09ICIuIiBlbHNlIFBhdGgoZXhlY3V0YWJsZSkubmFtZS5sb3dlcigpCi0gICAgICAgIGFyZ3NfbGlzdCA9IGxpc3QoYXJncykKLSAgICAgICAgY2FuZGlkYXRlcy5leHRlbmQoX3NoZWxsX3NlZ21l
