--- .github/pr2870-finalize/009.b64
@@ -1,1 +1,0 @@
-bnRfZmlsZV9vcGVyYW5kX3Rva2VucyhbbmFtZSwgKmFyZ3NfbGlzdF0pKQotICAgICAgICBpZiBuYW1lIGluIF9PVEhFUl9SRUFERVJTOgotICAgICAgICAgICAgY2FuZGlkYXRlcy5leHRlbmQoYXJnIGZvciBhcmcgaW4gYXJnc19saXN0IGlmIG5vdCBhcmcuc3RhcnRzd2l0aCgiLSIpKQotICAgICAgICBpZiBuYW1lIGluIHsic291cmNlIiwgIi4ifToKLSAgICAgICAgICAgIGNhbmRpZGF0ZXMuZXh0ZW5kKGFyZ3NfbGlzdFs6MV0pCi0gICAgICAgIGlmIG5hbWUgaW4geyJub2RlIiwgImJ1biIsICJydWJ5IiwgInBlcmwifSBvciBfcHl0aG9uX2V4ZWN1dGFibGUobmFtZSk6Ci0gICAgICAgICAgICBmb3IgaW5kZXgsIGFyZyBpbiBlbnVtZXJhdGUoYXJnc19saXN0WzotMV0pOgotICAgICAgICAgICAgICAgIGlmIGFyZyBpbiB7Ii1jIiwgIi1lIiwgIi0tZXZhbCIsICItcCIsICItLXByaW50In06Ci0gICAgICAgICAgICAgICAgICAgIGNhbmRpZGF0ZXMuZXh0ZW5kKF9saXRlcmFsX3JlYWRfcGF0aHMoYXJnc19saXN0W2luZGV4ICsgMV0pKQotICAgICAgICAgICAgICAgIGVsaWYgYXJnLnN0YXJ0c3dpdGgoKCItLWV2YWw9IiwgIi0tcHJpbnQ9IikpOgotICAgICAgICAgICAgICAgICAgICBjYW5kaWRhdGVzLmV4dGVuZChfbGl0ZXJhbF9yZWFkX3BhdGhzKGFyZy5zcGxpdCgiPSIsIDEpWzFdKSkKLSAgICBmb3IgcmVkaXJlY3QgaW4gY29tbWFuZC5yZWRpcmVjdHM6Ci0gICAgICAgIGlmIHJlZGlyZWN0Lm9wZXJhdG9yLmxzdHJpcCgiMDEyMzQ1Njc4OSIpIGluIHsiPCIsICI8PiJ9OgotICAgICAgICAgICAgY2FuZGlkYXRlcy5hcHBlbmQocmVkaXJlY3QudGFyZ2V0KQotICAgIHJldHVybiB0dXBsZSgKLSAgICAgICAgZGljdC5mcm9ta2V5cygKLSAgICAgICAgICAgIHBhdGggZm9yIHZhbHVlIGluIGNhbmRpZGF0ZXMgaWYgKHBhdGggOj0gX3NlbnNpdGl2ZV9wYXRoKHZhbHVlLCBjd2Q9Y3dkLCBob21lX2Rpcj1ob21lX2RpcikpIGlzIG5vdCBOb25lCi0gICAgICAgICkKLSAgICApCi0KLQotZGVmIF9zaGVsbF9jb21tYW5kX3N0cmluZyhleGVjdXRhYmxlOiBzdHIsIGFyZ3M6IHR1cGxlW3N0ciwgLi4uXSkgLT4gdHVwbGVbc3RyIHwgTm9uZSwgYm9vbF06Ci0gICAgIiIiUmV0dXJuIGEgbGl0ZXJhbCBzaGVsbCBjb21tYW5kIHN0cmluZyBhbmQgd2hldGhlciBjb21tYW5kLXN0cmluZyBtb2RlIHdhcyByZXF1ZXN0ZWQuIiIiCi0KLSAgICBuYW1lID0gIi4iIGlmIGV4ZWN1dGFibGUgPT0gIi4iIGVsc2UgUGF0aChleGVjdXRhYmxlKS5uYW1lLmxvd2VyKCkKLSAgICBpZiBuYW1lIG5vdCBpbiBfU0hFTExTOgotICAgICAgICByZXR1cm4gTm9uZSwgRmFsc2UKLSAgICBpbmRleCA9IDAKLSAgICB3aGlsZSBpbmRleCA8IGxlbihhcmdzKToKLSAgICAgICAgYXJnID0gYXJnc1tpbmRleF0KLSAgICAgICAgaWYgYXJnID09ICItLSI6Ci0gICAgICAgICAgICByZXR1cm4gTm9uZSwgRmFsc2UKLSAgICAgICAgaWYgYXJnID09ICItLWNvbW1hbmQiOgotICAgICAgICAgICAgcmV0dXJuIChhcmdzW2luZGV4ICsgMV0gaWYgaW5kZXggKyAxIDwgbGVuKGFyZ3MpIGVsc2UgTm9uZSksIFRydWUKLSAgICAgICAgaWYgYXJnLnN0YXJ0c3dpdGgoIi0tY29tbWFuZD0iKToKLSAgICAgICAgICAgIHJldHVybiBhcmcuc3BsaXQoIj0iLCAxKVsxXSwgVHJ1ZQotICAgICAgICBpZiBhcmcgPT0gIi1zIjoKLSAgICAgICAgICAgIHJldHVybiBOb25lLCBGYWxzZQotICAgICAgICBpZiBhcmcuc3RhcnRzd2l0aCgiLSIpIGFuZCBub3QgYXJnLnN0YXJ0c3dpdGgoIi0tIikgYW5kICJjIiBpbiBhcmdbMTpdOgotICAgICAgICAgICAgZnJvbSAuaW50ZXJwcmV0ZXJfb3B0aW9ucyBpbXBvcnQgc2hlbGxfaW50ZXJwcmV0ZXJfY29tbWFuZF9wYXlsb2FkCi0KLSAgICAgICAgICAgIHBhcnNlZCA9IHNoZWxsX2ludGVycHJldGVyX2NvbW1hbmRfcGF5bG9hZChbZXhlY3V0YWJsZSwgKmFyZ3NdLCAwKQotICAgICAgICAgICAgcmV0dXJuIChwYXJzZWQuc2NyaXB0X3RleHQgaWYgcGFyc2VkIGlzIG5vdCBOb25lIGVsc2UgTm9uZSksIFRydWUKLSAgICAgICAgaWYgYXJnIGluIHsiLW8iLCAiLU8iLCAiK28iLCAiK08iLCAiLS1yY2ZpbGUiLCAiLS1pbml0LWZpbGUifToKLSAgICAgICAgICAgIGluZGV4ICs9IDIKLSAgICAgICAgICAgIGNvbnRpbnVlCi0gICAgICAgIGlmIG5vdCBhcmcuc3RhcnRzd2l0aCgoIi0iLCAiKyIpKToKLSAgICAgICAgICAgIHJldHVybiBOb25lLCBGYWxzZQotICAgICAgICBpbmRleCArPSAxCi0gICAgcmV0dXJuIE5vbmUsIEZhbHNlCi0KLQotZGVmIF9zY3JpcHRfb3BlcmFuZChleGVjdXRhYmxlOiBzdHIsIGFyZ3M6IHR1cGxlW3N0ciwgLi4uXSkgLT4gdHVwbGVbc3RyLCBib29sXSB8IE5vbmU6Ci0gICAgIiIiUmV0dXJuIG9uZSBsaXRlcmFsIGxvY2FsIHNjcmlwdCBvcGVyYW5kIHRoYXQgY2FuIGJlIGluc3BlY3RlZCBzYWZlbHkuIiIiCi0KLSAgICBuYW1lID0gIi4iIGlmIGV4ZWN1dGFibGUgPT0gIi4iIGVsc2UgUGF0aChleGVjdXRhYmxlKS5uYW1lLmxvd2VyKCkKLSAgICBpc19zaGVsbCA9IG5hbWUgaW4gX1NIRUxMUwotICAgIGlzX2ludGVycHJldGVyID0gbmFtZSBpbiB7Im5vZGUiLCAicnVieSIsICJwZXJsIn0gb3IgX3B5dGhvbl9leGVjdXRhYmxlKG5hbWUpCi0gICAgaWYgaXNfc2hlbGwgb3IgaXNfaW50ZXJwcmV0ZXI6Ci0gICAgICAgIGluZGV4ID0gMAotICAgICAgICB3aGlsZSBpbmRleCA8IGxlbihhcmdzKToKLSAgICAgICAgICAgIGFyZyA9IGFyZ3NbaW5kZXhdCi0gICAgICAgICAgICBpZiAoCi0gICAgICAgICAgICAgICAgYXJnIGluIHsiLWMiLCAiLWUiLCAiLS1ldmFsIiwgIi1tIiwgIi0tY29tbWFuZCJ9Ci0gICAgICAgICAgICAgICAgb3IgKGlzX3NoZWxsIGFuZCBhcmcgPT0gIi1zIikKLSAgICAgICAgICAgICAgICBvciAoaXNfc2hlbGwgYW5kIGFyZy5zdGFydHN3aXRoKCItIikgYW5kIG5vdCBhcmcuc3RhcnRzd2l0aCgiLS0iKSBhbmQgImMiIGluIGFyZ1sx
