--- .github/pr2870-finalize/010.b64
@@ -1,1 +1,0 @@
-Ol0pCi0gICAgICAgICAgICApOgotICAgICAgICAgICAgICAgIHJldHVybiBOb25lCi0gICAgICAgICAgICBpZiBhcmcgPT0gIi0tIjoKLSAgICAgICAgICAgICAgICBpbmRleCArPSAxCi0gICAgICAgICAgICAgICAgYnJlYWsKLSAgICAgICAgICAgIGlmIGFyZyBpbiB7Ii1vIiwgIi1PIiwgIitvIiwgIitPIiwgIi0tcmNmaWxlIiwgIi0taW5pdC1maWxlIn06Ci0gICAgICAgICAgICAgICAgaW5kZXggKz0gMgotICAgICAgICAgICAgICAgIGNvbnRpbnVlCi0gICAgICAgICAgICBpZiBub3QgYXJnLnN0YXJ0c3dpdGgoKCItIiwgIisiKSk6Ci0gICAgICAgICAgICAgICAgYnJlYWsKLSAgICAgICAgICAgIGluZGV4ICs9IDEKLSAgICAgICAgaWYgaW5kZXggPCBsZW4oYXJncykgYW5kIGFyZ3NbaW5kZXhdICE9ICItIjoKLSAgICAgICAgICAgIHJldHVybiBhcmdzW2luZGV4XSwgaXNfc2hlbGwKLSAgICBpZiBuYW1lID09ICJidW4iIGFuZCBhcmdzIGFuZCBhcmdzWzBdLmVuZHN3aXRoKF9TQ1JJUFRfU1VGRklYRVMpOgotICAgICAgICByZXR1cm4gYXJnc1swXSwgRmFsc2UKLSAgICBpZiBleGVjdXRhYmxlLmVuZHN3aXRoKF9TQ1JJUFRfU1VGRklYRVMpIGFuZCAoIi8iIGluIGV4ZWN1dGFibGUgb3IgZXhlY3V0YWJsZS5zdGFydHN3aXRoKCIuIikpOgotICAgICAgICByZXR1cm4gZXhlY3V0YWJsZSwgZXhlY3V0YWJsZS5lbmRzd2l0aCgoIi5zaCIsICIuYmFzaCIsICIuenNoIiwgIi5rc2giLCAiLmZpc2giKSkKLSAgICByZXR1cm4gTm9uZQotCi0KLWRlZiBfa25vd25fcHl0aG9uX21vZHVsZV9sYXVuY2goZXhlY3V0YWJsZTogc3RyLCBhcmdzOiB0dXBsZVtzdHIsIC4uLl0sICosIGN3ZDogUGF0aCB8IE5vbmUpIC0+IGJvb2w6Ci0gICAgIiIiUmV0dXJuIFRydWUgZm9yIGEgdHJ1c3RlZCB0b29saW5nIG1vZHVsZSBoYW5kbGVkIGJ5IGRlZGljYXRlZCBwb2xpY3kuIiIiCi0KLSAgICBuYW1lID0gUGF0aChleGVjdXRhYmxlIG9yICIiKS5uYW1lLmxvd2VyKCkKLSAgICBpZiBub3QgX3B5dGhvbl9leGVjdXRhYmxlKG5hbWUpOgotICAgICAgICByZXR1cm4gRmFsc2UKLSAgICBmcm9tIC5zZWNyZXRfZmlsZV9yZXF1ZXN0X3NlcnZpY2VzLmNvbnN0YW50c19jb3JlIGltcG9ydCBfU0FGRV9QWVRIT05fTU9EVUxFX0NPTU1BTkRTCi0gICAgZnJvbSAuc2VjcmV0X2ZpbGVfcmVxdWVzdF9zZXJ2aWNlcy5pbnRlcnByZXRlcl9vYnNlcnZlcnMgaW1wb3J0IF9weXRob25fbW9kdWxlX21heV9iZV9zaGFkb3dlZAotICAgIGZyb20gLnNlY3JldF9maWxlX3JlcXVlc3Rfc2VydmljZXMucHl0ZXN0X2NvbmZpZ19zYWZldHkgaW1wb3J0IF9weXRob25fbW9kdWxlX3Jvb3RfZnJvbV9hcmdzCi0KLSAgICBtb2R1bGVfcm9vdCA9IF9weXRob25fbW9kdWxlX3Jvb3RfZnJvbV9hcmdzKGxpc3QoYXJncykpCi0gICAgcmV0dXJuIGJvb2wobW9kdWxlX3Jvb3QgaW4gX1NBRkVfUFlUSE9OX01PRFVMRV9DT01NQU5EUyBhbmQgbm90IF9weXRob25fbW9kdWxlX21heV9iZV9zaGFkb3dlZChtb2R1bGVfcm9vdCwgY3dkKSkKLQotCi1kZWYgX3B5dGhvbl9tb2R1bGVfbGF1bmNoKGV4ZWN1dGFibGU6IHN0ciwgYXJnczogdHVwbGVbc3RyLCAuLi5dLCAqLCBjd2Q6IFBhdGggfCBOb25lKSAtPiBib29sOgotICAgICIiIlJldHVybiBUcnVlIHdoZW4gUHl0aG9uIG1heSBpbXBvcnQgbXV0YWJsZSB3b3Jrc3BhY2UgY29kZSB2aWEgYGAtbWBgLiIiIgotCi0gICAgbmFtZSA9IFBhdGgoZXhlY3V0YWJsZSBvciAiIikubmFtZS5sb3dlcigpCi0gICAgaWYgbm90IF9weXRob25fZXhlY3V0YWJsZShuYW1lKToKLSAgICAgICAgcmV0dXJuIEZhbHNlCi0gICAgaWYgX2tub3duX3B5dGhvbl9tb2R1bGVfbGF1bmNoKGV4ZWN1dGFibGUsIGFyZ3MsIGN3ZD1jd2QpOgotICAgICAgICByZXR1cm4gRmFsc2UKLSAgICBmb3IgaW5kZXgsIGFyZyBpbiBlbnVtZXJhdGUoYXJncyk6Ci0gICAgICAgIGlmIGFyZyA9PSAiLS0iOgotICAgICAgICAgICAgcmV0dXJuIEZhbHNlCi0gICAgICAgIGlmIGFyZyA9PSAiLW0iIG9yIChhcmcuc3RhcnRzd2l0aCgiLW0iKSBhbmQgbGVuKGFyZykgPiAyKToKLSAgICAgICAgICAgIHJldHVybiBUcnVlCi0gICAgICAgIGlmIGFyZyBpbiB7Ii1XIiwgIi1YIn0gYW5kIGluZGV4ICsgMSA8IGxlbihhcmdzKToKLSAgICAgICAgICAgIGNvbnRpbnVlCi0gICAgICAgIGlmIG5vdCBhcmcuc3RhcnRzd2l0aCgiLSIpOgotICAgICAgICAgICAgcmV0dXJuIEZhbHNlCi0gICAgcmV0dXJuIEZhbHNlCi0KLQotZGVmIF9pbnRlcnByZXRlcl9pbmxpbmVfbGF1bmNoKGV4ZWN1dGFibGU6IHN0ciwgYXJnczogdHVwbGVbc3RyLCAuLi5dKSAtPiBib29sOgotICAgICIiIlJldHVybiBUcnVlIGZvciBsaXRlcmFsIGlubGluZS1jb2RlIG1vZGVzIHdob3NlIGZpbGUgcmVhZHMgd2VyZSBhbHJlYWR5IHNjYW5uZWQuIiIiCi0KLSAgICBuYW1lID0gUGF0aChleGVjdXRhYmxlIG9yICIiKS5uYW1lLmxvd2VyKCkKLSAgICBpZiBuYW1lIG5vdCBpbiB7Im5vZGUiLCAiYnVuIiwgInJ1YnkiLCAicGVybCJ9IGFuZCBub3QgX3B5dGhvbl9leGVjdXRhYmxlKG5hbWUpOgotICAgICAgICByZXR1cm4gRmFsc2UKLSAgICBmb3IgYXJnIGluIGFyZ3M6Ci0gICAgICAgIGlmIGFyZyA9PSAiLS0iOgotICAgICAgICAgICAgcmV0dXJuIEZhbHNlCi0gICAgICAgIGlmIGFyZyBpbiB7Ii0iLCAiLWMiLCAiLWUiLCAiLS1ldmFsIiwgIi1wIiwgIi0tcHJpbnQifSBvciBhcmcuc3RhcnRzd2l0aCgoIi0tZXZhbD0iLCAiLS1wcmludD0iKSk6Ci0gICAgICAgICAgICByZXR1cm4gVHJ1ZQotICAgICAgICBpZiBub3QgYXJnLnN0YXJ0c3dpdGgoIi0iKToKLSAgICAgICAgICAgIHJldHVybiBGYWxzZQotICAgIHJldHVybiBGYWxzZQotCi0KLWRlZiBfcGF0aF9xdWFsaWZpZWQoZXhlY3V0YWJsZTogc3RyKSAtPiBib29sOgotICAgIHJldHVybiBib29sKGV4ZWN1dGFibGUpIGFuZCAoIi8iIGluIGV4ZWN1dGFibGUgb3IgIlxcIiBpbiBleGVjdXRhYmxl
