--- .github/pr2870-finalize/011.b64
@@ -1,1 +1,0 @@
-IG9yIGV4ZWN1dGFibGUuc3RhcnRzd2l0aCgiLiIpKQotCi0KLWRlZiBfbG9jYWxfZXhlY3V0YWJsZV9vcGVyYW5kKAotICAgIGV4ZWN1dGFibGU6IHN0ciwKLSAgICAqLAotICAgIGN3ZDogUGF0aCB8IE5vbmUsCi0gICAgaG9tZV9kaXI6IFBhdGggfCBOb25lLAotICAgIHJvb3RzOiB0dXBsZVtQYXRoLCAuLi5dLAotKSAtPiBzdHIgfCBOb25lOgotICAgICIiIlJldHVybiBhIHBhdGgtcXVhbGlmaWVkIGV4ZWN1dGFibGUgb25seSB3aGVuIGl0IGlzIGluc2lkZSBhIGd1YXJkZWQgbG9jYWwgcm9vdC4iIiIKLQotICAgIGlmIG5vdCBfcGF0aF9xdWFsaWZpZWQoZXhlY3V0YWJsZSk6Ci0gICAgICAgIHJldHVybiBOb25lCi0gICAgbGV4aWNhbCA9IFBhdGgobm9ybWFsaXplX3BhdGgoZXhwYW5kX2hvbWUoZXhlY3V0YWJsZSwgaG9tZV9kaXIpLCBjd2QpKQotICAgIGlmIG5vdCBsZXhpY2FsLmlzX2Fic29sdXRlKCkgb3Igbm90IGFueShsZXhpY2FsLmlzX3JlbGF0aXZlX3RvKHJvb3QpIGZvciByb290IGluIHJvb3RzKToKLSAgICAgICAgcmV0dXJuIE5vbmUKLSAgICByZXR1cm4gZXhlY3V0YWJsZQotCi0KLWRlZiBfcGFyc2VfZXhlY3V0aW9uX3NlZ21lbnQoCi0gICAgZXhlY3V0aW9uOiBTaGVsbEV4ZWN1dGlvblNlZ21lbnQsCi0gICAgKiwKLSAgICByYXdfbW9kZWw6IENhbm9uaWNhbENvbW1hbmQsCi0gICAgcmF3X3NlZ21lbnQ6IENvbW1hbmRTZWdtZW50IHwgTm9uZSwKLSkgLT4gQ2Fub25pY2FsQ29tbWFuZCB8IE5vbmU6Ci0gICAgIiIiS2VlcCBzb3VyY2Ugc3ludGF4IGFuZCByZXF1aXJlIGFncmVlbWVudCB3aXRoIHRoZSBjd2QgbW9kZWwuCi0KLSAgICBSZS1xdW90aW5nIGRlY29kZWQgdG9rZW5zIGNoYW5nZXMgaW5wdXQgcmVkaXJlY3Rpb25zIGludG8gb3JkaW5hcnkgYXJndgotICAgIGFuZCBjYW4gZXJhc2UgY29tbWFuZC9leGVjIGFtYmlndWl0eS4gVGhlIHJhdyBjb21tYW5kIHBhcnNlciBkZWxpYmVyYXRlbHkKLSAgICBza2lwcyB0cmFuc3BhcmVudC13cmFwcGVyIG5vcm1hbGl6YXRpb24gZm9yIHRoaXMgaW5zcGVjdGlvbiBwYXRoLgotICAgICIiIgotCi0gICAgaWYgbm90IGV4ZWN1dGlvbi5jb21wbGV0ZSBvciBleGVjdXRpb24uZWZmZWN0aXZlX2N3ZCBpcyBOb25lIG9yIHJhd19zZWdtZW50IGlzIE5vbmU6Ci0gICAgICAgIHJldHVybiBOb25lCi0gICAgdHJ5OgotICAgICAgICBpZiBzcGxpdF9zaGVsbF90b2tlbnMocmF3X3NlZ21lbnQudGV4dCkgIT0gZXhlY3V0aW9uLnRva2VuczoKLSAgICAgICAgICAgIHJldHVybiBOb25lCi0gICAgZXhjZXB0IFZhbHVlRXJyb3I6Ci0gICAgICAgIHJldHVybiBOb25lCi0gICAgcmV0dXJuIHJlcGxhY2UoCi0gICAgICAgIHJhd19tb2RlbCwKLSAgICAgICAgc2VnbWVudHM9KHJhd19zZWdtZW50LCksCi0gICAgICAgIHJlZGlyZWN0cz10dXBsZSgKLSAgICAgICAgICAgIHJlZGlyZWN0Ci0gICAgICAgICAgICBmb3IgcmVkaXJlY3QgaW4gcmF3X21vZGVsLnJlZGlyZWN0cwotICAgICAgICAgICAgaWYgcmF3X3NlZ21lbnQuc3RhcnQgPD0gcmVkaXJlY3Quc3RhcnQgYW5kIHJlZGlyZWN0LmVuZCA8PSByYXdfc2VnbWVudC5lbmQKLSAgICAgICAgKSwKLSAgICAgICAgZW1iZWRkZWRfY29tbWFuZHM9KCksCi0gICAgKQotCi0KLWRlZiBfc2VnbWVudF9tYXlfdG91Y2hfbG9jYWxfZGF0YShleGVjdXRpb246IFNoZWxsRXhlY3V0aW9uU2VnbWVudCkgLT4gYm9vbDoKLSAgICAiIiJLZWVwIHVua25vd24gZmlsZS9jb2RlIGFjY2VzcyBjbG9zZWQgd2l0aG91dCB0cmVhdGluZyBzdGRvdXQgYXMgYSBmaWxlLiIiIgotCi0gICAgZnJvbSAuc2VjcmV0X2ZpbGVfcmVxdWVzdF9zZXJ2aWNlcy5sb2NhbF9yZWFkX29wZXJhbmRzIGltcG9ydCBfc2hlbGxfc2VnbWVudF9maWxlX29wZXJhbmRfdG9rZW5zCi0KLSAgICBpZiBub3QgZXhlY3V0aW9uLnRva2VuczoKLSAgICAgICAgcmV0dXJuIEZhbHNlCi0gICAgdG9rZW4gPSBleGVjdXRpb24udG9rZW5zWzBdCi0gICAgZXhlY3V0YWJsZSA9ICIuIiBpZiB0b2tlbiA9PSAiLiIgZWxzZSBQYXRoKHRva2VuKS5uYW1lLmxvd2VyKCkKLSAgICBhcmdzID0gbGlzdChleGVjdXRpb24udG9rZW5zWzE6XSkKLSAgICBoYXNfaW5wdXRfcmVkaXJlY3QgPSBhbnkocmUubWF0Y2gociJeXGQqPCg/ITwpIiwgaXRlbSkgZm9yIGl0ZW0gaW4gZXhlY3V0aW9uLnRva2VucykKLSAgICBpZiBoYXNfaW5wdXRfcmVkaXJlY3Qgb3IgX3NoZWxsX3NlZ21lbnRfZmlsZV9vcGVyYW5kX3Rva2VucyhbZXhlY3V0YWJsZSwgKmFyZ3NdKToKLSAgICAgICAgcmV0dXJuIFRydWUKLSAgICBpZiBleGVjdXRhYmxlIGluIF9PVEhFUl9SRUFERVJTOgotICAgICAgICByZXR1cm4gYW55KG5vdCBhcmcuc3RhcnRzd2l0aCgiLSIpIGZvciBhcmcgaW4gYXJncykKLSAgICByZXR1cm4gKAotICAgICAgICBleGVjdXRhYmxlIGluIHsqX1NIRUxMUywgImNvbW1hbmQiLCAiZXhlYyIsICJub2RlIiwgImJ1biIsICJydWJ5IiwgInBlcmwifQotICAgICAgICBvciBfcHl0aG9uX2V4ZWN1dGFibGUoZXhlY3V0YWJsZSkKLSAgICAgICAgb3IgX3BhdGhfcXVhbGlmaWVkKHRva2VuKQotICAgICkKLQotCi1kZWYgX2Zsb3dfb3BlcmF0b3JfYmVmb3JlKGV4ZWN1dGlvbjogU2hlbGxFeGVjdXRpb25TZWdtZW50KSAtPiBzdHIgfCBOb25lOgotICAgIHJldHVybiBuZXh0KCh0b2tlbiBmb3IgdG9rZW4gaW4gcmV2ZXJzZWQoZXhlY3V0aW9uLmNvbnRyb2xfYmVmb3JlKSBpZiB0b2tlbiBpbiBfRkxPV19PUEVSQVRPUlMpLCBOb25lKQotCi0KLWRlZiBfZmFpbGVkX2NkX3Nob3J0X2NpcmN1aXRfc3RhdGUoCi0gICAgZXhlY3V0aW9uOiBTaGVsbEV4ZWN1dGlvblNlZ21lbnQsCi0gICAgKiwKLSAgICBhY3RpdmU6IGJvb2wsCi0pIC0+IHR1cGxlW2Jvb2wsIGJvb2xdOgotICAgICIiIlJldHVybiB0aGUgZmFpbGVkLWNkIHN0YXRlIGFuZCB3aGV0aGVyIHRoaXMgc2VnbWVudCBpcyBwcm92YWJseSB1bnJlYWNoYWJsZS4iIiIKLQotICAgIG9wZXJhdG9y
