--- .github/pr2870-finalize/012.b64
@@ -1,1 +1,0 @@
-ID0gX2Zsb3dfb3BlcmF0b3JfYmVmb3JlKGV4ZWN1dGlvbikKLSAgICBpZiBvcGVyYXRvciBpbiB7Inx8IiwgIjsiLCAiJiJ9OgotICAgICAgICBhY3RpdmUgPSBGYWxzZQotICAgIGlmIGV4ZWN1dGlvbi5kaXJlY3Rvcnlfb3BlcmF0aW9uIGlzIG5vdCBOb25lIGFuZCBleGVjdXRpb24ucmVhc29uX2NvZGUgaW4gX1NIT1JUX0NJUkNVSVRJTkdfQ0RfRkFJTFVSRVM6Ci0gICAgICAgIHJldHVybiBUcnVlLCBGYWxzZQotICAgIHJldHVybiBhY3RpdmUsIGJvb2woYWN0aXZlIGFuZCBvcGVyYXRvciBpbiB7IiYmIiwgInwifSkKLQotCiBkZWYgYXNzZXNzX3NoZWxsX3JlYWRzKAogICAgIGNvbW1hbmRfdGV4dDogc3RyLAogICAgICosCkBAIC02NjMsNiArMjQ0LDExIEBAIGRlZiBhc3Nlc3Nfc2hlbGxfcmVhZHMoCiAgICAgICAgICAgICAgICAgY29udGludWUKICAgICAgICAgICAgIGlmIF9rbm93bl9weXRob25fbW9kdWxlX2xhdW5jaChleGVjdXRhYmxlLCBhcmd1bWVudHMsIGN3ZD1lZmZlY3RpdmVfY3dkKToKICAgICAgICAgICAgICAgICBjb250aW51ZQorICAgICAgICAgICAgaWYgX2ludGVycHJldGVyX3N0ZGluX2xhdW5jaChleGVjdXRhYmxlLCBhcmd1bWVudHMpOgorICAgICAgICAgICAgICAgIGlmIG5vdCBvd25lZF9oZXJlZG9jczoKKyAgICAgICAgICAgICAgICAgICAgcmVxdWVzdGVkID0gVHJ1ZQorICAgICAgICAgICAgICAgICAgICBpbmNvbXBsZXRlID0gVHJ1ZQorICAgICAgICAgICAgICAgIGNvbnRpbnVlCiAgICAgICAgICAgICBpZiBfaW50ZXJwcmV0ZXJfaW5saW5lX2xhdW5jaChleGVjdXRhYmxlLCBhcmd1bWVudHMpOgogICAgICAgICAgICAgICAgIGNvbnRpbnVlCiAgICAgICAgICAgICBpbnZvY2F0aW9uID0gX3NjcmlwdF9vcGVyYW5kKGV4ZWN1dGFibGUsIGFyZ3VtZW50cykKZGlmZiAtLWdpdCBhL3Rlc3RzL3Rlc3RfZ3VhcmRfc2hlbGxfc2VjcmV0X3JlYWRfZmxvb3JzLnB5IGIvdGVzdHMvdGVzdF9ndWFyZF9zaGVsbF9zZWNyZXRfcmVhZF9mbG9vcnMucHkKaW5kZXggMjhjNmQwYy4uY2NlMTI5YiAxMDA2NDQKLS0tIGEvdGVzdHMvdGVzdF9ndWFyZF9zaGVsbF9zZWNyZXRfcmVhZF9mbG9vcnMucHkKKysrIGIvdGVzdHMvdGVzdF9ndWFyZF9zaGVsbF9zZWNyZXRfcmVhZF9mbG9vcnMucHkKQEAgLTEzOCw2ICsxMzgsMjQgQEAgZGVmIHRlc3RfcHl0aG9uX2ZsYWdzX2JlZm9yZV9zY3JpcHRfZG9fbm90X2hpZGVfbG9jYWxfZXhlY3V0aW9uKHRtcF9wYXRoOiBQYXRoKQogICAgICAgICBhc3NlcnQgZXZhbHVhdGVfY29tbWFuZChjb21tYW5kLCBjd2Q9dG1wX3BhdGgsIGhvbWVfZGlyPXRtcF9wYXRoKS5taW5pbXVtX2FjdGlvbiA9PSAicmV2aWV3IgogCiAKK0BweXRlc3QubWFyay5wYXJhbWV0cml6ZSgKKyAgICAoImNvbW1hbmQiLCAiZmlsZW5hbWUiKSwKKyAgICAoCisgICAgICAgICgicHl0aG9uMyAtIDwgcGF5bG9hZC5weSIsICJwYXlsb2FkLnB5IiksCisgICAgICAgICgibm9kZSAtIDwgcGF5bG9hZC5qcyIsICJwYXlsb2FkLmpzIiksCisgICAgKSwKKykKK2RlZiB0ZXN0X2ludGVycHJldGVyX3N0ZGluX3JlcXVpcmVzX2xvY2FsX2V4ZWN1dGlvbl9yZXZpZXcoY29tbWFuZDogc3RyLCBmaWxlbmFtZTogc3RyLCB0bXBfcGF0aDogUGF0aCkgLT4gTm9uZToKKyAgICAodG1wX3BhdGggLyBmaWxlbmFtZSkud3JpdGVfdGV4dCgicHJpbnQoJ3N5bnRoZXRpYyBpbnB1dCcpXG4iLCBlbmNvZGluZz0idXRmLTgiKQorICAgIGFzc2Vzc21lbnQgPSBhc3Nlc3Nfc2hlbGxfcmVhZHMoY29tbWFuZCwgY3dkPXRtcF9wYXRoLCBob21lX2Rpcj10bXBfcGF0aCkKKyAgICBhc3NlcnQgYXNzZXNzbWVudC5zY3JpcHRfcmVxdWVzdGVkCisgICAgYXNzZXJ0IGFzc2Vzc21lbnQuaW5jb21wbGV0ZQorICAgIGFzc2VydCBhc3Nlc3NtZW50LnJlcXVpcmVzX3JldmlldworICAgIGV2YWx1YXRpb24gPSBldmFsdWF0ZV9jb21tYW5kKGNvbW1hbmQsIGN3ZD10bXBfcGF0aCwgaG9tZV9kaXI9dG1wX3BhdGgpCisgICAgYXNzZXJ0IGV2YWx1YXRpb24ubWluaW11bV9hY3Rpb24gPT0gInJldmlldyIKKyAgICBhc3NlcnQgZXZhbHVhdGlvbi5kZWNpc2lvbl9wbGFuZS5hY3Rpb24gPT0gInJlcXVpcmUtcmVhcHByb3ZhbCIKKworCiBkZWYgdGVzdF9weXRob25fbW9kdWxlX21vZGVfaXNfbXV0YWJsZV9sb2NhbF9leGVjdXRpb24odG1wX3BhdGg6IFBhdGgpIC0+IE5vbmU6CiAgICAgKHRtcF9wYXRoIC8gInJlYWRlci5weSIpLndyaXRlX3RleHQoJ29wZW4oIi5lbnYiKS5yZWFkKClcbicpCiAgICAgYXNzZXNzbWVudCA9IGFzc2Vzc19zaGVsbF9yZWFkcygicHl0aG9uMyAtSSAtbSByZWFkZXIiLCBjd2Q9dG1wX3BhdGgsIGhvbWVfZGlyPXRtcF9wYXRoKQo=
