+++ tests/fixtures/guard-command-corpus/decision-diff-report.json
@@ -136,7 +136,7 @@
       "source-5252fd793b49a537852d881d": "718973d25bf5d91943a27453c26ec505c00c0b89e827757e1ae6edf362955be5",
       "source-52870a8c6d89146820e8a067": "524d92e15a2e89c51cc1fcdb69c379208f2e5684e836cc885b7dc02e1d47cc90",
       "source-5384753b6f0132290e82a624": "23cdf9e4d27d0fe7c669d575108319121b02597ccafd9de77109f785800b3b49",
-      "source-5392a6fb1493c13aac80c872": "ba525f1a32b046a9937ef5e638a894a71e7473680a2d3ec5a9d61b0e8f26b5fd",
+      "source-5392a6fb1493c13aac80c872": "f92ef9b5657f4049444c9333b2296de4e2e09f35a0f47cb990561c07a537a98d",
       "source-53955d708b1f26505c7d4767": "db7bde370c9682b5c24e3fca3f266317660e3c966cedfdca548a6f1a524cca6d",
       "source-54b8965ce6adb239b24474a4": "282c3284ae3cec6a373c11f8b19694a7fad809bd59aa2fd633f0e4c3b41f5218",
       "source-54ca9f8f882d1002ba7d7bac": "47d191710b9a903f539375d25b3b356057d09480874f862ec676173727595ecb",
@@ -294,6 +294,7 @@
       "source-b970eabd271372ddfe730bf1": "d07b8da7b625a170acee771889ea340c049731b28c032b6179931de746adef5a",
       "source-ba3ba0028bdc49b9008e93d2": "ff785a7ea257c305fbd5d732bc1194d814c638e06d6816f25c4156148dd4e729",
       "source-ba4fee2dac2a5b0da5d61bdb": "b2434db33fbad299c552ab1a60923d451930cd4186b6d085f9fbcc7693aaee21",
+      "source-ba818a9c79b3640ab92f12ac": "3664c8849fbcff486ed171d4c636c1724895703a2a926c7f4befa88dbf78d958",
       "source-ba85b41a08ce4b871e40e7b7": "2d5ed4199691c6bee541d2328659ff00180214449d07137d436e28e128d41a26",
       "source-bbf1900d9722bd89273df0a9": "bcd3c7a6bc266408b38e3b9de145fa98c6db549d82e9c79b637085e97a74eda1",
       "source-bc036bfe24f968f06679d17c": "9771e902dc9bca564b291d2e80990a66464e1f4e732f0e5b5bd521600c7e00c4",
